### 2018-10-26 #1051 @whedon
[PRE REVIEW]: morphMan: Automated manipulation of vascular geometries
**Submitting author:** @aslakbergersen (<a href="http://orcid.org/0000-0001-5063-3680">Aslak Bergersen</a>)
**Repository:** <a href="https://github.com/KVSlab/morphMan" target ="_blank">https://github.com/KVSlab/morphMan</a>
**Version:** v0.1
