### 2018-10-26 e6cbe66
Fixed reference typo. Used markdown for images. (HTML did not compile in TeX-version)
