### 2018-10-26 8efb1ad
Re-placed figure1
