### 2018-10-26 #1051 @rlizzo
Thanks for the note @aslakbergersen. Count me in. From a quick look this seems like an interesting project/paper.
