### 2018-10-26 #1051 @rlizzo
As a matter of disclosure to the editor @trallard. I am one of the lead developers of [VMTK](www.github.com/vmtk/vmtk), which this project uses extensively; in addition I was an author of VMTK's [accepted JOSS paper](http://joss.theoj.org/papers/10.21105/joss.00745#), which was for the version of the VMTK library this project requires.

However, as I do not have any prior relationship (professional or personal) with the listed authors or their employer - Simula Research Laboratory - I do not believe this to be a [conflict of interest](https://github.com/openjournals/joss/blob/master/COI.md). However, it should be publicly noted for the integrity of the open review process.
