### 2018-11-02 #1051 @aslakbergersen
Hey @trallard, @brainstorm is a potential reviewer with a perfect background for giving a good review on this project. He reviewed the vmtk paper in JOSS.
