### 2018-11-04 #1051 @arfon
@brainstorm `assign` clobbers all existing assignments. Sorry the commands aren't a little clearer...
