### 2018-11-04 #1065 @whedon
[REVIEW]: morphMan: Automated manipulation of vascular geometries
**Submitting author:** @aslakbergersen (<a href="http://orcid.org/0000-0001-5063-3680">Aslak Bergersen</a>)
**Repository:** <a href="https://github.com/KVSlab/morphMan" target ="_blank">https://github.com/KVSlab/morphMan</a>
**Version:** v0.2
