### 2018-11-04 #1065 @aslakbergersen
@brainstorm thank you for agreeing to review morphMan. I'm not sure I understand what you are asking. There are 13 references in the paper.md, but the bibliography (paper.bib) is located [here.](https://github.com/KVSlab/morphMan/blob/master/paper/paper.bib) I hope that answered your question.
