### 2018-11-05 #1065 @brainstorm
My bad, I was reading the [rendered version](https://github.com/KVSlab/morphMan/blob/master/paper/paper.md) and I saw that section empty, but they are indeed in the .bib as they should, ticking off that box ;)
