### 2018-11-05 KVSlab/morphMan#11 @rlizzo
JOSS Review - Blocking Issue - LICENSE File
This is a Blocking issue for the project's JOSS Submission: https://github.com/openjournals/joss-reviews/issues/1065

Two things:

1) Remove the `.md` from the `LICENSE.md` file. It is important to have a pure plain text license file (which it not markdown formatted) for automated LICENSE finders to include appropriately.

2) Please list the applicable licenses for VMTK / VTK / ITK (and other dependent projects), basically, copy paste (and modify where appropriate) the [VMTK LICENSE file](https://github.com/vmtk/vmtk/blob/master/LICENSE)
