### 2018-11-06 KVSlab/morphMan#12 @rlizzo
JOSS Review - Blocking Issue - Paper Audience
This is a blocking issue for the JOSS Review at: https://github.com/openjournals/joss-reviews/issues/1065

The `Software Paper` Section states:

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

Please specify who the target audience is for the tool. Is this for researchers, clinicians, students? What level of programming background does the user need to have to use the tool?
