### 2018-11-06 KVSlab/morphMan#13 @rlizzo
JOSS Review - Blocking Issue - Community Guidelines
This is a blocking issue for the JOSS Review: https://github.com/openjournals/joss-reviews/issues/1065

The JOSS Submission Requires:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Please add documentation about how to set up an environment for development and/or guidelines on contributions. Preferably this would be in a separate `CONTRIBUTING.md` file.
