### 2018-11-06 KVSlab/morphMan#14 @rlizzo
JOSS Review - Installation Method
This is for the JOSS Review: https://github.com/openjournals/joss-reviews/issues/1065.

As this is a pure python package, it seems fairly simply to add a [`setup.py` script](https://docs.python.org/3.6/distutils/setupscript.html) which would install this package as a system python package. The only additions required would be either:
1) specifying the appropriate `entry_point` for each script in `setup.py`
2) add a simple CLI wrapper for the scripts via [argparse](https://docs.python.org/3.6/library/argparse.html), or [click](https://click.palletsprojects.com/en/7.x/). Which would accomplish the same as (1) in a much cleaner (and probably easier) way.

Pure python packages can be trivial converted to conda packages via [Conda Build](https://conda.io/docs/user-guide/tasks/build-packages/index.html) and then uploaded to [Anaconda Cloud](https://anaconda.org) where VMTK and other dependencies would be downloaded and linked "auto-magically" via Anaconda.

**I'm not going to mark this as a blocking issue**, since installation is easily enough performed, and the current solution satisfies the minimum JOSS requirements.

However, I would request feedback on why this hasn't been prioritized & included from the start?   It's certainly possible that I'm missing something important about the code and/or project structure which makes this a more complex task than it initially appears? My concern stems from the fact that anyone who finds the package needs to not only know the field, but also be comfortable with Git/Github, and needs to know how to work with python in a more complex way than pulling up any terminal in any directory and typing the `morphman foo` command. *It's OK to require that knowledge, but it should probably be stated as "prerequisite knowledge" somewhere in the documentation or JOSS paper.*
