### 2018-11-06 1c00ac3
[Issue #11] Removed LICENSE file extension and added licenses for morphMan dependencies.
