### 2018-11-06 03f5778
[Issue #11] Included morphMan in applicable licenses.
