### 2018-11-06 a796203
[Issue #11] Typo fix.
