### 2018-11-06 483c0cd
[Issue #12] Expanded paper.md with target audience, application, and main prerequisites.
