+++ paper/paper.md
@@ -42,7 +42,7 @@ oversimplified results. To use realistic geometries we could instead correlate t
 stresses based on hundreds of patient-specific models, but this is very labor intensive and
 error-prone [@valen2018real, @berg2018multiple].
 
-Our goal was to take the best from both approaches and create a tool which could parameterize patient-specific
+Our goal was to take the best from both approaches and create a tool which could parametrize patient-specific
 geometries to mimic the natural variability of morphological features in the population. We here present a framework,
 *morphMan*, that allows for *objective*, *reproducible*, and *automatic* virtual manipulation of tubular structures
 here exemplified with application to the cerebrovasculature.
@@ -58,17 +58,22 @@ diagram are not connected, and therefore easier to manipulate. As a result, only
 [@Piccinelli2011]; moreover, a subset of the algorithms are presented in [@Bergersen2016] and [@kjeldsberg2018investigating].
 
 In *morphMan v0.1* you can alter cross-sectional area, bifurcation angles, 
-overall curvature in a segment, and the shape of specific bends. For each
+overall curvature in a segment, and the shape of specific bends. 
+For each
 category, there is a wide range of options, thus providing the users with many degrees of
-freedom for manipulating the geometries. Shown in Figure 2 is an example of rotating the
-branches in a bifurcation 'up' and 'down'. *morphMan* is easily expandable for specialized manipulations.
+freedom for manipulating the geometries. 
+Shown in Figure 2 is an example of rotating the
+branches in a bifurcation 'up' and 'down'. 
 
 ![manipulation ICA bifurcation](./figure2.png)\
 
 **Figure 2:** 
    Output from *morphMan* for manipulating the bifurcation angles.
 
-*morphMan* is a general tool for manipulating tubular objects. In the context of flow in the cardiovascular system,
+Although the framework is applicable by anyone familiar with the Python programming language and basic use of Git/Github, the intended audience of *morphMan* is researchers, particularly within the field of computational biomechanics. Our goal is to provide researchers, and other individuals, with a set of tools to investigate the impact of altering morphological features in patient-specific geometries. For instance, by combining morphMan with a computational fluid mechanics solver, it can be used to objectively and reproducibly investigate how cross-sectional area, and overall curvature impacts the local hemodynamics. Previously, one would have to investigate tens, if not hundreds, of models to correlate changes in specific morphological parameters to changes in local hemodynamic. Now, one can keep all other morphological parameters fixed while investigating only the impact of one change. Thus, morphMan opens a wide range of problems which previously was practically infeasible to study.
+
+To summarize, *morphMan* is a general tool for automated manipulation of tubular objects, easily expandable for specialized manipulations.
+In the context of flow in the cardiovascular system,
 *morphMan* opens new lines of investigation for unraveling the coupling between
 morphology and the computed local hemodynamics.
 
