### 2018-11-06 707cbf2
[Issue #13] Created guidelines for contributing (draft), including a code of conduct.
