+++ CONTRIBUTING.md
@@ -1,0 +1,127 @@
+# Contributing to morphMan
+
+When contributing to this repository, please first discuss the change you wish to make via issue,
+email, or any other method with the owners of this repository before making a change. 
+
+
+
+### Testing 
+
+Please provide unit tests for the  new code you create, testing the main functionality or feature to be submitted. We can always use more test coverage! 
+
+### Submitting changes
+
+In order to submitt you changes, please send a [GitHub Pull Request to morphMan](https://github.com/KVSlab/morphMan/pull/new/master) with a clear list of what you've done (read more about [pull requests](http://help.github.com/pull-requests/)). Please follow our coding conventions (below) and make sure all of your commits are atomic (one feature per commit).
+
+Always write a clear commit message when submitting your changes. One-line messages are fine for small changes, but bigger changes should look like this:
+
+    > $ git commit -m "A brief summary of the commit
+    > 
+    > A paragraph describing what changed and its impact."
+
+### Coding conventions
+
+#### Formatting 
+* Avoid inline comments.
+* Break long lines after 120 characters.
+* Delete trailing whitespace.
+* Don't include spaces after `(`, `[` or before `]`, `)`.
+* Don't misspell.
+* Use 4 space indentation.
+* Use an empty line between methods.
+* Use spaces around operators, except for unary operators, such as `!`.
+* Use spaces after commas, after colons and semicolons, around `{` and before
+  `}`.
+
+#### Naming
+
+* Avoid abbreviations.
+* Use snake case for variables and methods.
+* Use camel case for classes.
+* Name variables, methods, and classes to reveal intent.
+
+#### Organization
+
+* Order methods so that caller methods are earlier in the file than the methods
+  they call.
+* Place methods receiving command line arguments at the bottom of the file, but above the top-level script environment check.
+* Separate local and global imports of modules.
+
+
+### Code of Conduct
+
+### Our Pledge
+
+In the interest of fostering an open and welcoming environment, we as
+contributors and maintainers pledge to making participation in our project and
+our community a harassment-free experience for everyone, regardless of age, body
+size, disability, ethnicity, sex characteristics, gender identity and expression,
+level of experience, education, socio-economic status, nationality, personal
+appearance, race, religion, or sexual identity and orientation.
+
+### Our Standards
+
+Examples of behavior that contributes to creating a positive environment
+include:
+
+* Using welcoming and inclusive language
+* Being respectful of differing viewpoints and experiences
+* Gracefully accepting constructive criticism
+* Focusing on what is best for the community
+* Showing empathy towards other community members
+
+Examples of unacceptable behavior by participants include:
+
+* The use of sexualized language or imagery and unwelcome sexual attention or
+  advances
+* Trolling, insulting/derogatory comments, and personal or political attacks
+* Public or private harassment
+* Publishing others' private information, such as a physical or electronic
+  address, without explicit permission
+* Other conduct which could reasonably be considered inappropriate in a
+  professional setting
+
+### Our Responsibilities
+
+Project maintainers are responsible for clarifying the standards of acceptable
+behavior and are expected to take appropriate and fair corrective action in
+response to any instances of unacceptable behavior.
+
+Project maintainers have the right and responsibility to remove, edit, or
+reject comments, commits, code, wiki edits, issues, and other contributions
+that are not aligned to this Code of Conduct, or to ban temporarily or
+permanently any contributor for other behaviors that they deem inappropriate,
+threatening, offensive, or harmful.
+
+### Scope
+
+This Code of Conduct applies both within project spaces and in public spaces
+when an individual is representing the project or its community. Examples of
+representing a project or community include using an official project e-mail
+address, posting via an official social media account, or acting as an appointed
[…]
