### 2018-11-06 53b7fd7
Update paper.md
