### 2018-11-07 KVSlab/morphMan#11 @aslakbergersen
@rlizzo Thank you for pointing this out. We were actually a bit unsure on how to properly use the LICENSE. If you are pleased with the corrections we can close this issue.

Just to check, are the smaller version of the LICENCE on top of each file sufficient?
