### 2018-11-07 2703a13
[Issue #13] Fixed typos.
