### 2018-11-07 KVSlab/morphMan#13 @aslakbergersen
@rlizzo, thank you for the comment. We have now added a `CONTRIBUTING.md` file, see [here.](https://github.com/KVSlab/morphMan/blob/master/CONTRIBUTING.md) If you find the changes to be satisfactory we can close this issue.
