### 2018-11-07 cf0bc89
Basic python module setup.
