+++ test/test_landmarking.py
@@ -9,8 +9,8 @@
 import sys
 from os import path
 relative_path = path.dirname(path.abspath(__file__))
-sys.path.insert(0, path.join(relative_path, '..', 'src'))
-sys.path.insert(0, "../src")
+sys.path.insert(0, path.join(relative_path, '..', 'morphman'))
+sys.path.insert(0, "../morphman")
 
 import pytest
 from common import *
