+++ build/lib/morphMan/common.py
@@ -1,0 +1,4194 @@
+##   Copyright (c) Aslak W. Bergersen, Henrik A. Kjeldsberg. All rights reserved.
+##   See LICENSE file for details.
+
+##      This software is distributed WITHOUT ANY WARRANTY; without even 
+##      the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR 
+##      PURPOSE.  See the above copyright notices for more information.
+
+import math
+import sys
+from os import path, makedirs
+
+import numpy as np
+import numpy.linalg as la
+import vtk
+from scipy.interpolate import splrep, splev
+from scipy.signal import resample
+from vmtk import vtkvmtk, vmtkscripts
+from vtk.util import numpy_support
+
+# Local import
+from vmtkpointselector import *
+
+# Global array names
+radiusArrayName = 'MaximumInscribedSphereRadius'
+parallelTransportNormalsArrayName = 'ParallelTransportNormals'
+groupIDsArrayName = "GroupIds"
+abscissasArrayName = 'Abscissas'
+clippingArrayName = 'ClippingArray'
+branchClippingArrayName = 'BranchClippingArray'
+distanceToTubeArrayName = 'DistanceToTubeFunction'
+closedArrayName = 'ClosedSection'
+eikonalSolutionArrayName = 'EikonalSolutionArray'
+edgeArrayName = 'EdgeArray'
+edgePCoordArrayName = 'EdgePCoordArray'
+costFunctionArrayName = 'CostFunctionArray'
+
+# Options not available from commandline
+divergingRatioToSpacingTolerance = 2.0
+interpolationHalfSize = 3
+voronoiCoreCutOffThreshold = 0.75
+numberOfSplineAnalyzedPoints = 40
+phiValues = [float(i) for i in range(2, 43, 2)]
+thetaStep = 2.0
+
+
+def read_polydata(filename, datatype=None):
+    """
+    Load the given file, and return a vtkPolyData object for it.
+
+    Args:
+        filename (str): Path to input file.
+        datatype (str): Additional parameter for vtkIdList objects.
+
+    Returns:
+        polyData (vtkSTL/vtkPolyData/vtkXMLStructured/
+                    vtkXMLRectilinear/vtkXMLPolydata/vtkXMLUnstructured/
+                    vtkXMLImage/Tecplot): Output data.
+    """
+
+    # Check if file exists
+    if not path.exists(filename):
+        raise RuntimeError("Could not find file: %s" % filename)
+
+    # Check filename format
+    fileType = filename.split(".")[-1]
+    if fileType == '':
+        raise RuntimeError('The file does not have an extension')
+
+    # Get reader
+    if fileType == 'stl':
+        reader = vtk.vtkSTLReader()
+        reader.MergingOn()
+    elif fileType == 'vtk':
+        reader = vtk.vtkPolyDataReader()
+    elif fileType == 'vtp':
+        reader = vtk.vtkXMLPolyDataReader()
+    elif fileType == 'vts':
+        reader = vtk.vtkXMinkorporereLStructuredGridReader()
+    elif fileType == 'vtr':
+        reader = vtk.vtkXMLRectilinearGridReader()
+    elif fileType == 'vtu':
+        reader = vtk.vtkXMLUnstructuredGridReader()
+    elif fileType == "vti":
+        reader = vtk.vtkXMLImageDataReader()
+    elif fileType == "np" and datatype == "vtkIdList":
+        result = np.load(filename).astype(np.int)
+        id_list = vtk.vtkIdList()
+        id_list.SetNumberOfIds(result.shape[0])
+        for i in range(result.shape[0]):
+            id_list.SetId(i, result[i])
+        return id_list
[…]
