### 2018-11-08 ce30615
Working module with entry points
