+++ build/lib/morphMan/manipulate_curvature.py
@@ -1,339 +1,0 @@
-##   Copyright (c) Aslak W. Bergersen, Henrik A. Kjeldsberg. All rights reserved.
-##   See LICENSE file for details.
-
-##      This software is distributed WITHOUT ANY WARRANTY; without even 
-##      the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR 
-##      PURPOSE.  See the above copyright notices for more information.
-
-from argparse import ArgumentParser, RawDescriptionHelpFormatter
-
-# Local import
-from common import *
-from argparse_common import *
-
-
-def curvature_variations(input_filepath, smooth, smooth_factor, smooth_factor_line, iterations,
-                         smooth_line, output_filepath, poly_ball_size, region_of_interest,
-                         region_points, resampling_step, no_smooth, no_smooth_point):
-    """
-    Create a sharper or smoother version of the input geometry,
-    determined by a smoothed version of the siphon centerline.
-
-    Args:
-        input_filepath (str): Path to input surface.
-        output_filepath (str): Path to output the manipulated surface.
-        smooth (bool): Smooth the Voronoi diagram.
-        smooth_line (bool): Smooth centerline if True, anti-smooth if False.
-        smooth_factor (float): Smoothing factor used for Voronoi diagram smoothing.
-        smooth_factor_line (float): Smoothing factor used for centerline smoothing.
-        iterations (int): Smoothing iterations of centerline.
-        region_of_interest (str): Method for setting the region of interest ['manual' | 'commandline' | 'first_line']
-        region_points (list): If region_of_interest is 'commandline', this a flatten list of the start and endpoint
-        poly_ball_size (list): Resolution of polyballs used to create surface.
-        resampling_step (float): Resampling length for centerline resampling.
-        no_smooth (bool): True of part of the model is not to be smoothed.
-        no_smooth_point (ndarray): Point which is untouched by smoothing.
-    """
-    # Input filenames
-    base_path = get_path_names(input_filepath)
-
-    # Centerlines
-    direction = "smoothed" if smooth_line else "extended"
-    centerlines_path = base_path + "_centerline.vtp"
-    centerline_smooth_path = base_path + "_centerline_smoothed.vtp"
-    new_centerlines_path = base_path + "_centerline_new_%s.vtp" % direction
-
-    # Clean and capp / uncapp surface
-    surface, capped_surface = prepare_surface(base_path, input_filepath)
-
-    # Voronoi diagrams filenames
-    voronoi_remaining_path = base_path + "_voronoi_curvature_remaining.vtp"
-    voronoi_region_path = base_path + "_voronoi_region.vtp"
-    voronoi_diverging_path = base_path + "_voronoi_diverging.vtp"
-
-    # Compute centerlines
-    inlet, outlets = get_centers(surface, base_path)
-    centerlines, voronoi, pole_ids = compute_centerlines(inlet, outlets, centerlines_path,
-                                                         capped_surface, resampling=resampling_step,
-                                                         smooth=False, base_path=base_path)
-    if smooth:
-        voronoi = prepare_voronoi_diagram(capped_surface, centerlines, base_path,
-                                          smooth, smooth_factor, no_smooth,
-                                          no_smooth_point, voronoi, pole_ids)
-    # Get region of interest
-    _, _, _, region_points = get_line_to_change(capped_surface, centerlines,
-                                                region_of_interest, "variation", region_points, 0)
-    region_points = [[region_points[3 * i], region_points[3 * i + 1], region_points[3 * i + 2]]
-                     for i in range(len(region_points) // 3)]
-
-    # Set and get clipping points, centerlines and diverging centerlines
[…]
