+++ setup.py
@@ -3,30 +3,47 @@ import setuptools
 with open("README.md", "r") as fh:
     long_description = fh.read()
 
+
+DEPENDENCIES = [] #'scipy', 'numpy', 'vtk']
+TEST_DEPENDENCIES = ['pytest']
+
+VERSION = "0.1"
+URL = "https://github.com/KVSlab/morphman.git"
+
 setuptools.setup(
     name="morphman",
-    version="0.1",
+    version=VERSION,
     license="GPL",
     author="Aslak W. Bergersen, Henrik A. Kjeldsberg",
     author_email="henrik.kjeldsberg@live.no",
+    url=URL,
+    project_urls={
+        "Documentation": "https://morphman.readthedocs.io/",
+        "Source Code": URL,
+    },
     description="morphman - morphlogical manipulation",
     long_description=long_description,
     long_description_content_type="text/markdown",
-    url="https://github.com/KVSlab/morphman.git",
-    packages=setuptools.find_packages(exclude=["test"]),
-    project_urls={
-        "Documentation": "https://morphman.readthedocs.io/",
-        "Source Code": "https://github.com/KVSlab/morphman",
-    },
-    install_requires=[
-        'scipy', 'numpy', 'vtk'
-    ],
-    dependency_links=[
-        'https://github.com/vmtk/vmtk'
-    ],
+
+    # Dependencies
+    install_requires=DEPENDENCIES,
+    tests_require=TEST_DEPENDENCIES,
+    dependency_links=['https://github.com/vmtk/vmtk'],
+
     classifiers=[
         'Intended Audience :: Developers',
         'Intended Audience :: Science/Research',
         "Programming Language :: Python :: 3",
     ],
+
+    packages=["morphman"],
+    package_dir = {"morphman": "morphman"},
+    #setuptools.find_packages(exclude=["test"]),
+
+    entry_points = {'console_scripts':
+                        ['morphman-area=morphman.manipulate_area:main_area',
+                         'morphman-bend=morphman.manipulate_bend:main_bend',
+                         'morphman-bifurcation=morphman.manipulate_angle:main_bifurcation',
+                         'morphman-curvature=morphman.manipulate_curvature:main_curvature']
+                   }
 )
