+++ build/lib/morphMan/manipulate_bifurcation.py
@@ -1,619 +1,0 @@
-##   Copyright (c) Aslak W. Bergersen, Henrik A. Kjeldsberg. All rights reserved.
-##   See LICENSE file for details.
-
-##      This software is distributed WITHOUT ANY WARRANTY; without even 
-##      the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR 
-##      PURPOSE.  See the above copyright notices for more information.
-
-from argparse import ArgumentParser, RawDescriptionHelpFormatter
-
-from argparse_common import *
-# Local import
-from common import *
-
-
-def rotate_branches(input_filepath, output_filepath, smooth, smooth_factor, angle,
-                    keep_fixed_1, keep_fixed_2, bif, lower, no_smooth, no_smooth_point,
-                    poly_ball_size, cylinder_factor, resampling_step,
-                    region_of_interest, region_points):
-    """
-    Objective rotation of daughter branches, by rotating
-    centerlines and Voronoi diagram about the bifurcation center.
-    The implementation is an extension of the original method
-    presented by Ford et al. (2009), for aneurysm removal,
-    which introduces the possibility to rotate the
-    daughter branches a given angle.
-    Includes the option to rotate only one of the daughter branches.
-
-    Args:
-        input_filepath (str): Path to input surface.
-        output_filepath (str): Path to output surface.
-        smooth (bool): Determine if the voronoi diagram should be smoothed.
-        smooth_factor (float): Smoothing factor used for voronoi diagram smoothing.
-        angle (float): Angle which daughter branches are moved, in radians.
-        keep_fixed_1 (bool): Leaves first branch untouched if True.
-        keep_fixed_2 (bool): Leaves second branch untouched if True.
-        bif (bool): Interpolates bifurcation is True.
-        lower (bool): Interpolates a lowered line through the bifurcation if True.
-        cylinder_factor(float): Factor for choosing the smaller cylinder during Voronoi interpolation.
-        resampling_step (float): Resampling step used to resample centerlines.
-        no_smooth (bool): True of part of the model is not to be smoothed.
-        no_smooth_point (ndarray): Point which is untouched by smoothing.
-        region_of_interest (str): Method for setting the region of interest ['manual' | 'commandline' ]
-        region_points (list): If region_of_interest is 'commandline', this a flatten list of the start and endpoint
-        poly_ball_size (list): Resolution of polyballs used to create surface.
-    """
-    # Filenames
-    base_path = get_path_names(input_filepath)
-
-    # Output filepaths
-    # Centerliens
-    centerline_par_path = base_path + "_centerline_par.vtp"
-    centerline_bif_path = base_path + "_centerline_bif.vtp"
-    centerline_clipped_path = base_path + "_centerline_clipped_ang.vtp"
-    centerline_clipped_bif_path = base_path + "_centerline_clipped_bif_ang.vtp"
-    centerline_new_path = base_path + "_centerline_interpolated_ang.vtp"
-    centerline_new_bif_path = base_path + "_centerline_interpolated_bif_ang.vtp"
-    centerline_new_bif_lower_path = base_path + "_centerline_interpolated_bif_lower_ang.vtp"
-    centerline_relevant_outlets_path = base_path + "_centerline_relevant_outlets.vtp"
-    centerline_rotated_path = base_path + "_centerline_rotated_ang.vtp"
-    centerline_rotated_bif_path = base_path + "_centerline_rotated_bif_ang.vtp"
[…]
