+++ test/test_manipulate_curvature.py
@@ -5,18 +5,12 @@
 ##      the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR 
 ##      PURPOSE.  See the above copyright notices for more information.
 
-import sys
-from os import path
-relative_path = path.dirname(path.abspath(__file__))
-sys.path.insert(0, path.join(relative_path, '..', 'morphman'))
-sys.path.insert(0, "../morphman")
-
 import pytest
 import numpy as np
-from .fixtures import common_input
-from manipulate_curvature import curvature_variations
-from common import get_path_names, extract_single_line, read_polydata, get_locator
-from estimate_alpha_and_beta import discrete_geometry
+
+from fixtures import common_input
+from morphman import manipulate_curvature, get_path_names, extract_single_line, \
+                     read_polydata, get_locator, discrete_geometry
 
 
 @pytest.mark.parametrize("smooth_line", [True, False])
