+++ morphman/__init__.py
@@ -1,13 +1,5 @@
-#name = "morphman"
-
 # Main features
 from .manipulate_area import manipulate_area, main_area
 from .manipulate_bifurcation import manipulate_bifurcation, main_bifurcation
 from .manipulate_bend import manipulate_bend, main_bend
 from .manipulate_curvature import manipulate_curvature, main_curvature
-
-#import .vmtkpointselector
-#import .estimate_alpha_and_beta
-from .automated_landmarking import automated_landmarking
-
-#import .common
