+++ test/test_landmarking.py
@@ -8,9 +8,9 @@
 
 import pytest
 
-from morphman import get_path_names, extract_ica_centerline, landmarking_bogunovic, \
-                     landmarking_piccinelli
-from fixtures import common_input
+from morphman.common import get_path_names, extract_ica_centerline
+from morphman.misc import landmarking_bogunovic, landmarking_piccinelli
+from .fixtures import common_input
 
 
 @pytest.mark.parametrize("algorithm", ["bogunovic", "piccinelli"])
