+++ morphman/manipulate_area.py
@@ -9,8 +9,7 @@ from scipy.ndimage.filters import gaussian_filter
 from argparse import ArgumentParser, RawDescriptionHelpFormatter
 
 # Local import
-from .common import *
-from .argparse_common import *
+from morphman.common import *
 
 
 def manipulate_area(input_filepath, method, smooth, smooth_factor, no_smooth,
