### 2018-11-08 b79249b
Update travis to install the package
