+++ .travis.yml
@@ -38,7 +38,9 @@ install:
   - sed -i -e 's/len(self.SourcePoints)\/3/len\(self.SourcePoints\)\/\/3/g' ../../../miniconda/envs/morphman/lib/python3.6/site-packages/vmtk/vmtkcenterlines.py
   - sed -i -e 's/len(self.TargetPoints)\/3/len\(self.TargetPoints\)\/\/3/g' ../../../miniconda/envs/morphman/lib/python3.6/site-packages/vmtk/vmtkcenterlines.py
 
+  # Install morphman
+  - python setup.py install
+
 script:
-  # Your test script goes here
   - cd test
   - pytest .
