### 2018-11-09 046aad7
Added build for AppVeyor
