### 2018-11-13 24e75dd
Updated documentation with usage through python module.
