+++ setup.py
@@ -3,8 +3,7 @@ import setuptools
 with open("README.md", "r") as fh:
     long_description = fh.read()
 
-
-DEPENDENCIES = [] #'scipy', 'numpy', 'vtk']
+DEPENDENCIES = ['scipy', 'numpy', 'vtk']
 TEST_DEPENDENCIES = ['pytest']
 
 VERSION = "0.1"
@@ -35,16 +34,14 @@ setuptools.setup(
         'Intended Audience :: Science/Research',
         "Programming Language :: Python :: 3",
     ],
-
     packages=["morphman",
               "morphman.common",
               "morphman.misc"],
-    package_dir = {"morphman": "morphman"},
-
-    entry_points = {'console_scripts':
-                        ['morphman-area=morphman.manipulate_area:main_area',
-                         'morphman-bend=morphman.manipulate_bend:main_bend',
-                         'morphman-bifurcation=morphman.manipulate_angle:main_bifurcation',
-                         'morphman-curvature=morphman.manipulate_curvature:main_curvature']
-                   }
+    package_dir={"morphman": "morphman"},
+    entry_points={'console_scripts':
+                      ['morphman-area=morphman.manipulate_area:main_area',
+                       'morphman-bend=morphman.manipulate_bend:main_bend',
+                       'morphman-bifurcation=morphman.manipulate_angle:main_bifurcation',
+                       'morphman-curvature=morphman.manipulate_curvature:main_curvature']
+                  }
 )
