### 2018-11-13 0a08e4f
Conda-build setup with req. scripts
