### 2018-11-20 95bdfb8
Working conda build script. Upated setup.py
