+++ morphman/manipulate_bend.py
@@ -82,7 +82,7 @@ def manipulate_bend(input_filepath, output_filepath, smooth, smooth_factor, regi
                                 " landmarking with automated_landmarking.py first.") % point_path)
         region_points = np.loadtxt(point_path)
     else:
-        _, _, _, region_points = get_line_to_change(capped_surface, centerlines,
+        _, _, _, region_points, _ = get_line_to_change(capped_surface, centerlines,
                                                     region_of_interest, "bend", region_points, 0)
         region_points = [[region_points[3 * i], region_points[3 * i + 1], region_points[3 * i + 2]]
                          for i in range(len(region_points) // 3)]
