+++ conda-build/conda_build_config.yaml
@@ -1,0 +1,6 @@
+python:
+  - 3.5.1
+  - 3.6.1
+
+vtk:
+  - 8.1.0
