### 2018-11-20 7de5fd9
Update CI scripts with conda build
