### 2018-11-20 f8a8a32
Re-added vmtkcenterlines fix
