### 2018-11-20 49e0387
Updated manipulate_area
