### 2018-11-20 91514b1
Updated CI scripts.
