+++ .travis.yml
@@ -30,7 +30,7 @@ install:
   # Set up environment
   - conda create -n morphman -c morphman -c vmtk python=3.6 morphman pytest
   # - conda create -q -n morphman -c vmtk python=3.6 itk vtk vmtk scipy numpy pytest
-  - source activate morphman_env
+  - source activate morphman
 
   # Useful for debugging any issues with conda
   - conda info -a
