### 2018-11-20 a2636fd
Update CI scripts
