+++ .travis.yml
@@ -26,10 +26,12 @@ install:
 
   # Set conda parameters
   - conda config --set always_yes yes --set changeps1 no
+  - conda config --add channels vmtk
  
   # Set up environment
-  - conda create -n morphman -c morphman -c vmtk python=3.6 morphman pytest
-  # - conda create -q -n morphman -c vmtk python=3.6 itk vtk vmtk scipy numpy pytest
+  - conda create -n morphman -c morphman python=3.6 morphman pytest
+
+  # Activate environment
   - source activate morphman
 
   # Useful for debugging any issues with conda
