+++ appveyor.yml
@@ -40,6 +40,7 @@ install:
   # Conda setup
   - "%CMD_IN_ENV% conda config --add channels vmtk"
   - "%CMD_IN_ENV% conda config --add channels morphman"
+  - "%CMD_IN_ENV% conda config --set always_yes yes"
 
   # Useful for debugging any issues with conda
   - "%CMD_IN_ENV% conda info -a"
@@ -52,7 +53,6 @@ install:
 
   # Install morphman
   - "%CMD_IN_ENV% conda install morphman"
-
 
   # Search and replace in vmtkcenterlines.py for / to //
   - sed -i 's/len(self.SourcePoints)\/3/len\(self.SourcePoints\)\/\/3/g' ..\\..\\Python36_64\\envs\\morphman_env\\lib\\site-packages\\vmtk\\vmtkcenterlines.py
