### 2018-11-20 07afee3
Edit in appveyor.yml
