### 2018-11-20 e4339ab
Fixed MISR issue in manipulate_area
