### 2018-11-21 38d0a3b
Added python 3.5 to CI script
