### 2018-11-21 52dfa7e
building in test script
