### 2018-11-21 cb6d822
Update numpy dep. in conda
