+++ conda-build/meta.yaml
@@ -16,8 +16,8 @@ requirements:
   run:
     - python {{ python }}
     - setuptools
-    - scipy >=1.0.0
-    - numpy >=1.13
+    - scipy 1.1.0
+    - numpy <=1.13
     - vtk {{ vtk }}
     - vmtk
 
