### 2018-11-21 c47de2e
Added dependencies to appveyor.
