+++ appveyor.yml
@@ -70,11 +70,11 @@ install:
   #- sed -i 's/len(self.TargetPoints)\/3/len\(self.TargetPoints\)\/\/3/g' ..\\..\\Python36_64\\envs\\morphman_env\\lib\\site-packages\\vmtk\\vmtkcenterlines.py
 
   # Workaround for downloading on windows, as bitsadmin is not in sync on AppVeyor
-  - cd .\\test
-  - ps: Start-FileDownload 'http://ecm2.mathcs.emory.edu/aneuriskdata/download/C0001/C0001_models.tar.gz'
-  - tar --force-local -zxvf C0001_models.tar.gz
-  - del /f C0001_models.tar.gz
-  - cd ..
+  #- cd .\\test
+  #- ps: Start-FileDownload 'http://ecm2.mathcs.emory.edu/aneuriskdata/download/C0001/C0001_models.tar.gz'
+  #- tar --force-local -zxvf C0001_models.tar.gz
+  #- del /f C0001_models.tar.gz
+  #- cd ..
   
 #build_script:
   # Build the compiled extension
@@ -82,6 +82,11 @@ install:
 build: false
 
 test_script:
+  - cd .\\test
+  - ps: Start-FileDownload 'http://ecm2.mathcs.emory.edu/aneuriskdata/download/C0001/C0001_models.tar.gz'
+  - tar --force-local -zxvf C0001_models.tar.gz
+  - del /f C0001_models.tar.gz
+  - cd ..
   - "%CMD_IN_ENV% conda update -q conda"
   - "%CMD_IN_ENV% conda config --set always_yes yes"
   - "%CMD_IN_ENV% conda config --add channels vmtk"
