### 2018-11-22 1475d1b
Update appveyor.yml
Navigate in windows
