+++ appveyor.yml
@@ -82,11 +82,21 @@ install:
 build: false
 
 test_script:
-  - cd .\\test
+  - dir
+  - cd \\projects\\morphman-s1s38\\test
+  - dir
   - ps: Start-FileDownload 'http://ecm2.mathcs.emory.edu/aneuriskdata/download/C0001/C0001_models.tar.gz'
   - tar --force-local -zxvf C0001_models.tar.gz
   - del /f C0001_models.tar.gz
-  - cd ..
+  
+  
+  # Search and replace in vmtkcenterlines.py for / to //
+  - sed -i 's/len(self.SourcePoints)\/3/len\(self.SourcePoints\)\/\/3/g' \\Python36_64\lib\site-packages\vmtk\vmtkcenterlines.py
+  - sed -i 's/len(self.TargetPoints)\/3/len\(self.TargetPoints\)\/\/3/g' \\Python36_64\lib\site-packages\vmtk\vmtkcenterlines.py
+
+  - sed -i 's/len(self.SourcePoints)\/3/len\(self.SourcePoints\)\/\/3/g' \\Python35_64\lib\site-packages\vmtk\vmtkcenterlines.py
+  - sed -i 's/len(self.TargetPoints)\/3/len\(self.TargetPoints\)\/\/3/g' \\Python35_64\lib\site-packages\vmtk\vmtkcenterlines.py
+  
   - "%CMD_IN_ENV% conda update -q conda"
   - "%CMD_IN_ENV% conda config --set always_yes yes"
   - "%CMD_IN_ENV% conda config --add channels vmtk"
