+++ appveyor.yml
@@ -83,19 +83,20 @@ build: false
 
 test_script:
   - dir
-  - cd \\projects\\morphman-s1s38\\test
+  - cd\
+  - cd .\\projects\\morphman-s1s38\\test
   - dir
   - ps: Start-FileDownload 'http://ecm2.mathcs.emory.edu/aneuriskdata/download/C0001/C0001_models.tar.gz'
   - tar --force-local -zxvf C0001_models.tar.gz
   - del /f C0001_models.tar.gz
-  
+  - cd\
   
   # Search and replace in vmtkcenterlines.py for / to //
-  - sed -i 's/len(self.SourcePoints)\/3/len\(self.SourcePoints\)\/\/3/g' \\Python36_64\lib\site-packages\vmtk\vmtkcenterlines.py
-  - sed -i 's/len(self.TargetPoints)\/3/len\(self.TargetPoints\)\/\/3/g' \\Python36_64\lib\site-packages\vmtk\vmtkcenterlines.py
+  - sed -i 's/len(self.SourcePoints)\/3/len\(self.SourcePoints\)\/\/3/g' .\\Python36_64\lib\site-packages\vmtk\vmtkcenterlines.py
+  - sed -i 's/len(self.TargetPoints)\/3/len\(self.TargetPoints\)\/\/3/g' .\\Python36_64\lib\site-packages\vmtk\vmtkcenterlines.py
 
-  - sed -i 's/len(self.SourcePoints)\/3/len\(self.SourcePoints\)\/\/3/g' \\Python35_64\lib\site-packages\vmtk\vmtkcenterlines.py
-  - sed -i 's/len(self.TargetPoints)\/3/len\(self.TargetPoints\)\/\/3/g' \\Python35_64\lib\site-packages\vmtk\vmtkcenterlines.py
+  - sed -i 's/len(self.SourcePoints)\/3/len\(self.SourcePoints\)\/\/3/g' .\\Python35_64\lib\site-packages\vmtk\vmtkcenterlines.py
+  - sed -i 's/len(self.TargetPoints)\/3/len\(self.TargetPoints\)\/\/3/g' .\\Python35_64\lib\site-packages\vmtk\vmtkcenterlines.py
   
   - "%CMD_IN_ENV% conda update -q conda"
   - "%CMD_IN_ENV% conda config --set always_yes yes"
