+++ appveyor.yml
@@ -45,11 +45,6 @@ install:
   - "SET PATH=%PYTHON%;%PYTHON%\\Scripts;%PATH%"
   - "set PATH=%MINICONDA%;%MINICONDA%\\Scripts;%PATH%"
   
-  - "%CMD_IN_ENV% conda update -q conda"
-  - "%CMD_IN_ENV% conda config --set always_yes yes"
-  - "%CMD_IN_ENV% conda config --add channels vmtk"
-  - "%CMD_IN_ENV% conda config --add channels morphman"
-  - "%CMD_IN_ENV% conda config --show"
   # Conda setup
   #- "%CMD_IN_ENV% conda config --add channels vmtk"
   #- "%CMD_IN_ENV% conda config --add channels morphman"
@@ -88,21 +83,24 @@ install:
 build: false
 
 test_script:
-  - dir
-  - cd\
-  - cd .\\projects\\morphman-s1s38\\test
-  - dir
-  - ps: Start-FileDownload 'http://ecm2.mathcs.emory.edu/aneuriskdata/download/C0001/C0001_models.tar.gz'
-  - tar --force-local -zxvf C0001_models.tar.gz
-  - del /f C0001_models.tar.gz
-  - cd\
-  - dir
+  #- dir
+  #- cd\
+  #- cd .\\projects\\morphman-s1s38\\test
+  #- dir
+  #- ps: Start-FileDownload 'http://ecm2.mathcs.emory.edu/aneuriskdata/download/C0001/C0001_models.tar.gz'
+  #- tar --force-local -zxvf C0001_models.tar.gz
+  #- del /f C0001_models.tar.gz
+  #- cd\
+  #- dir
   
   # Search and replace in vmtkcenterlines.py for / to //
   #- sed -i 's/len(self.SourcePoints)\/3/len\(self.SourcePoints\)\/\/3/g' .\\Python%CONDA_PY%_64\lib\site-packages\vmtk\vmtkcenterlines.py
   #- sed -i 's/len(self.TargetPoints)\/3/len\(self.TargetPoints\)\/\/3/g' .\\Python%CONDA_PY%_64\lib\site-packages\vmtk\vmtkcenterlines.py
-  
-
+  - "%CMD_IN_ENV% conda update -q conda"
+  - "%CMD_IN_ENV% conda config --set always_yes yes"
+  - "%CMD_IN_ENV% conda config --add channels vmtk"
+  - "%CMD_IN_ENV% conda config --add channels morphman"
+  - "%CMD_IN_ENV% conda config --show"
   - "%CMD_IN_ENV% conda install pytest vtk vmtk scipy numpy"
   - "%CMD_IN_ENV% conda build --python %APP_CONDA_PY% ."
   - "%CMD_IN_ENV% cd .\\test"
