+++ appveyor.yml
@@ -96,6 +96,8 @@ test_script:
   # Search and replace in vmtkcenterlines.py for / to //
   #- sed -i 's/len(self.SourcePoints)\/3/len\(self.SourcePoints\)\/\/3/g' .\\Python%CONDA_PY%_64\lib\site-packages\vmtk\vmtkcenterlines.py
   #- sed -i 's/len(self.TargetPoints)\/3/len\(self.TargetPoints\)\/\/3/g' .\\Python%CONDA_PY%_64\lib\site-packages\vmtk\vmtkcenterlines.py
+  
+  # Start building
   - "%CMD_IN_ENV% conda update -q conda"
   - "%CMD_IN_ENV% conda config --set always_yes yes"
   - "%CMD_IN_ENV% conda config --add channels vmtk"
@@ -103,5 +105,10 @@ test_script:
   - "%CMD_IN_ENV% conda config --show"
   - "%CMD_IN_ENV% conda install pytest vtk vmtk scipy numpy"
   - "%CMD_IN_ENV% conda build --python %APP_CONDA_PY% ."
+  
+  # Start testing
+  - "%CMD_IN_ENV% ps: Start-FileDownload 'http://ecm2.mathcs.emory.edu/aneuriskdata/download/C0001/C0001_models.tar.gz'"
+  - "%CMD_IN_ENV% tar --force-local -zxvf C0001_models.tar.gz"
+  - "%CMD_IN_ENV% del /f C0001_models.tar.gz"
   - "%CMD_IN_ENV% cd .\\test"
   - "%CMD_IN_ENV% pytest ."
