+++ appveyor.yml
@@ -107,8 +107,8 @@ test_script:
   - "%CMD_IN_ENV% conda build --python %APP_CONDA_PY% ."
   
   # Start testing
-  - "%CMD_IN_ENV% ps: Start-FileDownload 'http://ecm2.mathcs.emory.edu/aneuriskdata/download/C0001/C0001_models.tar.gz'"
-  - "%CMD_IN_ENV% tar --force-local -zxvf C0001_models.tar.gz"
-  - "%CMD_IN_ENV% del /f C0001_models.tar.gz"
+  - ps: Start-FileDownload 'http://ecm2.mathcs.emory.edu/aneuriskdata/download/C0001/C0001_models.tar.gz'
+  - tar --force-local -zxvf C0001_models.tar.gz
+  - del /f C0001_models.tar.gz
   - "%CMD_IN_ENV% cd .\\test"
   - "%CMD_IN_ENV% pytest ."
