+++ appveyor.yml
@@ -110,5 +110,5 @@ test_script:
   - ps: Start-FileDownload 'http://ecm2.mathcs.emory.edu/aneuriskdata/download/C0001/C0001_models.tar.gz'
   - tar --force-local -zxvf C0001_models.tar.gz
   - del /f C0001_models.tar.gz
-  - "%CMD_IN_ENV% cd .\\test"
-  - "%CMD_IN_ENV% pytest ."
+  - cd .\\test #"%CMD_IN_ENV% cd .\\test"
+  - pytest . #"%CMD_IN_ENV% pytest ."
