### 2018-11-22 d55c224
Build only in appyveyor.
