+++ appveyor.yml
@@ -44,57 +44,12 @@ install:
   - powershell .\\continuous-integration\\appveyor\\install.ps1
   - "SET PATH=%PYTHON%;%PYTHON%\\Scripts;%PATH%"
   - "set PATH=%MINICONDA%;%MINICONDA%\\Scripts;%PATH%"
-  
-  # Conda setup
-  #- "%CMD_IN_ENV% conda config --add channels vmtk"
-  #- "%CMD_IN_ENV% conda config --add channels morphman"
-  #- "%CMD_IN_ENV% conda config --set always_yes yes"
-
-  # Useful for debugging any issues with conda
-  #- "%CMD_IN_ENV% conda info -a"
-
-  # Update conda
-  #- "%CMD_IN_ENV% conda update -q conda"
-
-  # Create environment for morphman
-  #- "%CMD_IN_ENV% conda create -n morphman python=%PYTHON_VERSION% morphman vmtk"
-
-  # Activate environment
-  #- activate morphman
-
-  # Install morphman
-  #- conda install pytest
-
-  # Search and replace in vmtkcenterlines.py for / to //
-  #- sed -i 's/len(self.SourcePoints)\/3/len\(self.SourcePoints\)\/\/3/g' ..\\..\\Python36_64\\envs\\morphman_env\\lib\\site-packages\\vmtk\\vmtkcenterlines.py
-  #- sed -i 's/len(self.TargetPoints)\/3/len\(self.TargetPoints\)\/\/3/g' ..\\..\\Python36_64\\envs\\morphman_env\\lib\\site-packages\\vmtk\\vmtkcenterlines.py
-
-  # Workaround for downloading on windows, as bitsadmin is not in sync on AppVeyor
-  #- cd .\\test
-  #- ps: Start-FileDownload 'http://ecm2.mathcs.emory.edu/aneuriskdata/download/C0001/C0001_models.tar.gz'
-  #- tar --force-local -zxvf C0001_models.tar.gz
-  #- del /f C0001_models.tar.gz
-  #- cd ..
-  
-#build_script:
-  # Build the compiled extension
-  # - "%CMD_IN_ENV% python setup.py build"
 
 build: false
 
 test_script:
-  #- dir
-  #- cd\
-  #- cd .\\projects\\morphman-s1s38\\test
-  #- dir
-  #- ps: Start-FileDownload 'http://ecm2.mathcs.emory.edu/aneuriskdata/download/C0001/C0001_models.tar.gz'
-  #- tar --force-local -zxvf C0001_models.tar.gz
-  #- del /f C0001_models.tar.gz
-  #- cd\
-  #- dir
-  
-  
   # Start building
+  - dir
   - "%CMD_IN_ENV% conda update -q conda"
   - "%CMD_IN_ENV% conda config --set always_yes yes"
   - "%CMD_IN_ENV% conda config --add channels vmtk"
@@ -106,12 +61,12 @@ test_script:
   # Search and replace in vmtkcenterlines.py for / to //
   #- sed -i 's/len(self.SourcePoints)\/3/len\(self.SourcePoints\)\/\/3/g' .\\Python%CONDA_PY%_64\lib\site-packages\vmtk\vmtkcenterlines.py
   #- sed -i 's/len(self.TargetPoints)\/3/len\(self.TargetPoints\)\/\/3/g' .\\Python%CONDA_PY%_64\lib\site-packages\vmtk\vmtkcenterlines.py
-  
+
   # Start testing
-  - cd .\\test #"%CMD_IN_ENV% cd .\\test"
-  - dir
[…]
