+++ appveyor.yml
@@ -60,7 +60,7 @@ test_script:
   
   # Search and replace in vmtkcenterlines.py for / to //
   - dir
-  - "%CMD_IN_ENV% dir vmtkcenterlines.py /s /p"
+  - "%CMD_IN_ENV% dir /s vmtkcenterlines.py"
   - '%CMD_IN_ENV% sed -i "s/len(self.SourcePoints)\/3/len\(self.SourcePoints\)\/\/3/g" .\\Python%CONDA_PY%_64\\lib\\site-packages\\vmtk\\vmtkcenterlines.py'
   - '%CMD_IN_ENV% sed -i "s/len(self.TargetPoints)\/3/len\(self.TargetPoints\)\/\/3/g" .\\Python%CONDA_PY%_64\\lib\\site-packages\\vmtk\\vmtkcenterlines.py'
 
