+++ appveyor.yml
@@ -60,6 +60,13 @@ test_script:
   
   # Search and replace in vmtkcenterlines.py for / to //
   - dir
+  - cd Python%CONDA_PY%_64
+  - dir
+  - cd Library
+  - dir
+  - cd bin
+  - dir
+
   - "%CMD_IN_ENV% dir /s vmtkcenterlines.py"
   - '%CMD_IN_ENV% sed -i "s/len(self.SourcePoints)\/3/len\(self.SourcePoints\)\/\/3/g" .\\Python%CONDA_PY%_64\\lib\\site-packages\\vmtk\\vmtkcenterlines.py'
   - '%CMD_IN_ENV% sed -i "s/len(self.TargetPoints)\/3/len\(self.TargetPoints\)\/\/3/g" .\\Python%CONDA_PY%_64\\lib\\site-packages\\vmtk\\vmtkcenterlines.py'
