+++ appveyor.yml
@@ -59,12 +59,12 @@ test_script:
   - "%CMD_IN_ENV% conda build --python %APP_CONDA_PY% ."
   
   # Search and replace in vmtkcenterlines.py for / to //
-  - "%CMD_IN_ENV% dir"
-  - "%CMD_IN_ENV% cd Python%CONDA_PY%_64"
-  - "%CMD_IN_ENV% dir"
-  - "%CMD_IN_ENV% cd Library"
-  - "%CMD_IN_ENV% dir"
-  - "%CMD_IN_ENV% cd bin"
+  #- "%CMD_IN_ENV% dir"
+  #- "%CMD_IN_ENV% cd Python%CONDA_PY%_64"
+  #- "%CMD_IN_ENV% dir"
+  #- "%CMD_IN_ENV% cd Library"
+  #- "%CMD_IN_ENV% dir"
+  #- "%CMD_IN_ENV% cd bin"
 
   - '%CMD_IN_ENV% sed -i "s/len(self.SourcePoints)\/3/len\(self.SourcePoints\)\/\/3/g" ..\\..\\Python%CONDA_PY%_64\\lib\\site-packages\\vmtk\\vmtkcenterlines.py'
   - '%CMD_IN_ENV% sed -i "s/len(self.TargetPoints)\/3/len\(self.TargetPoints\)\/\/3/g" ..\\..\\Python%CONDA_PY%_64\\lib\\site-packages\\vmtk\\vmtkcenterlines.py'
