+++ conda-build/meta.yaml
@@ -11,13 +11,12 @@ build:
 requirements:
   build:
     - python
-    - setuptools
+    - setuptools             # [win]
 
   run:
     - python {{ python }}
-    - setuptools
-    - scipy #{{ scipy }}
-    - numpy #<=1.13
+    - scipy
+    - numpy <=1.13
     - vtk {{ vtk }}
     - vmtk
 
@@ -26,6 +25,9 @@ test:
     - pytest
   source_files:
     - ./test
+  commands:
+    - pytest ./test/
+
 
 about:
   home: https://github.com/KVSlab/morphMan
