### 2018-11-25 07ec614
Removed manual test in conda
