+++ conda-build/meta.yaml
@@ -25,8 +25,6 @@ test:
     - pytest
   source_files:
     - ./test
-  commands:
-    - pytest ./test/
 
 
 about:
