### 2018-11-26 ff2b6a0
Conda install working on win 64
