### 2018-11-26 #1065 @trallard
Hi @aslakbergersen  I can see there are still some open issues and unchecked items on the reviewer's lists so I wanted to check if everything was in order.
