### 2018-11-26 a9f47b9
Updated documentation with conda installation.
