### 2018-11-26 a7d91bd
Refining documentation
