### 2018-11-26 33b2c14
Updated .travis and installation to include vmtk channel
