### 2018-11-26 2f242f2
Typo .travis.yml
