### 2018-11-26 a6c62da
[Issue #14] Refactoring in conda build for all platforms.
