+++ conda-build/conda_build_config.yaml
@@ -4,6 +4,3 @@ python:
 
 vtk:
   - 8.1.0
-
-scipy:
-  - 1.1.0
