### 2018-11-26 KVSlab/morphMan#14 @brainstorm
Following up on @rlizzo comments, I think it would be even easier (from a software packaging/repo perspective), to add morphman on conda-forge:

https://conda-forge.org/#contribute

There's a lot of already packaged scientific software you can take as a reference and quickly get yours included in the pool.
