### 2018-11-27 KVSlab/morphMan#14 @rlizzo
> Following up on @rlizzo comments, I think it would be even easier (from a software packaging/repo perspective), to add morphman on conda-forge:
>
> https://conda-forge.org/#contribute
>
> There's a lot of already packaged scientific software you can take as a reference and quickly get yours included in the pool.

Totally Agreed! Moving to conda forge is on the todo list for me for VMTK as well. If you wanted to go that route, I could move over the build in the next few days so that dependencies aren't an issue?
