### 2018-11-27 KVSlab/morphMan#12 @rlizzo
Closing issue due to completion of tasks. thank you!
