### 2018-11-27 KVSlab/morphMan#11 @rlizzo
Hi @aslakbergersen, sorry for the late reply, yes it is sufficient. Closing due to completion of task! Thank you!
