### 2018-11-27 67c0cee
Small updates to documentation
