### 2018-11-27 KVSlab/morphMan#14 @aslakbergersen
Thank you for your comments @brainstorm and @rlizzo. We completely agree and have we now restructured morphMan to be a package which can be installed from conda. We have also added entry_points in the build script such that one can execute morphman-[area / bifurcation / bend / curvature]. We have updated the installation and tutorials accordingly. See pull request #19. We hope you will find the changes to be an improvement.

Given the current build, it should be very straightforward to move over to conda-forge once vmtk is there as well.
