### 2018-11-27 #1065 @aslakbergersen
Hi @trallard, thank you for checking in. After comments from both reviewers, we have just changed the installation method, meeting the comments in KVSlab/morphMan#14.

@brainstorm and @rlizzo is there anything else we should address?
