### 2018-12-10 #1065 @trallard
I can see that both reviewers left the following point of the checklist unchecked:
>  Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

@brainstorm @rlizzo can you advise on this? I saw there is a read the docs website but I am unsure if you both deemed that more work would need to be done at an API level
