### 2018-12-11 fb8196a
Update installation.rst
Add workaround for vmtk problem to installation instructions.
