+++ docs/source/installation.rst
@@ -16,14 +16,15 @@ The general dependencies of morphMan are
 * VTK 8.1.0
 * Numpy <= 1.13
 * SciPy 1.1.0
+* Python (2.7 or >=3.5)
 
 Basic Installation
 ==================
 We recommend that you can install morphMan through Anaconda.
 First, install Anaconda or Miniconda (preferably the python 3.6 version).
-Then execute the following command in the terminal window::
+Then execute the following command in a terminal window::
 
-  conda create -n your_environment -c vmtk -c morphman morphman
+  $ conda create -n your_environment -c vmtk -c morphman morphman
 
 You can then activate your environment by running ``source activate your_environment``.
 Now you are all set, and can start using morphMan. morphMan can be accessed by opening a Python console
@@ -38,15 +39,21 @@ Alternatively you can use one of the four main methods of manipulation directly 
 
 followed by the command line arguments for the selected method. A detailed explanation for usage of morphMan is described in :ref:`getting_started`.
 
+.. WARNING:: The vmtk version 1.4, the one currently distributed with anaconda, has a python3 bug in `vmtkcenterlines`. As a workaround you have to change this file. To find out where it is located please execute::
+  
+    $ which vmtkcenterlines
+    /Users/[Name]/miniconda3/envs/[your_environment]/bin/vmtkcenterlines
+    $ python -V
+    Python 3.6.2 :: Continuum Analytics, Inc.
+  
+  Now copy the  path up until ``[your_environment]`` and add ``lib/python3.6/site-packages/vmtk/vmtkcenterlines.py``. Please change the path seperation symbol to match your operating system and change ``python3.6`` to the python version you are using. Using this path you can run the two following lines::
+
+    $ sed -i -e 's/len(self.SourcePoints)\/3/len\(self.SourcePoints\)\/\/3/g' /Users/[Name]/miniconda3/envs/[your_environment]/lib/python3.6/site-packages/vmtk/vmtkcenterlines.py
+    $ sed -i -e 's/len(self.TargetPoints)\/3/len\(self.TargetPoints\)\/\/3/g' /Users/[Name]/miniconda3/envs/[your_environment]/lib/python3.6/site-packages/vmtk/vmtkcenterlines.py
+
+
 Development version
 ===================
-Requirements
-~~~~~~~~~~~~
-The latest development version of morphMan may be acquired through GitHub.
-In order to successfully build and use morphMan, the following additional software needs to be installed on your system:
-
-* Git (>=1.6)
-* Python (>=2.5)
 
 Downloading
 ~~~~~~~~~~~
@@ -56,7 +63,7 @@ Make sure Git (>=1.6) is installed, which is needed to clone the repository.
 To clone the morphMan repository, navigate to the directory where you wish
 morphMan to be stored, type the following command, and press Enter::
 
-    git clone https://github.com/KVSlab/morphMan
+    $ git clone https://github.com/KVSlab/morphMan
 
 After the source distribution has been downloaded, all the files required will be located
 in the newly created ``morphMan`` folder.
@@ -64,7 +71,7 @@ in the newly created ``morphMan`` folder.
 Building
 ~~~~~~~~
 In order to build and install morphMan, navigate into the ``morphMan`` folder, where a ``setup.py``
-file will be located. Building and installation of morphMan can be performed by simply running the
-following command from the terminal window::
+file will be located. First, make sure that all dependencies are installed. Then, building and installation of morphMan
+can be performed by simply running the following command from the terminal window::
 
-    python setup.py install
+    $ python setup.py install
