### 2018-12-11 #1065 @aslakbergersen
@brainstorm We would prefer to have morphman on conda-forge, however, since we rely on VMTK, which is not on conda-forge (yet), we instead created your own channel, for now, cf. [conda-forge dependencies](https://conda-forge.org/docs/recipe.html#avoid-dependencies-outside-of-conda-forge).

@rlizzo You mentioned that vmtk eventually will be moved over to conda-forge, how is that work going, and is there something we could contribute with?

@brainstorm, regarding running the test locally. Thank you for pointing this out, because this is a critical bug. In vmtk version 1.4 there is a python3 bug in how the centerlines are created, as a workaround, we modify the vmtk installation in both TravisCI and AppVayor, cf. [this line for travis](https://github.com/KVSlab/morphMan/blob/a3c6a18293603c449c6c23ba7334df347038960a/.travis.yml#L42) and [this line for AppVayor](https://github.com/KVSlab/morphMan/blob/a3c6a18293603c449c6c23ba7334df347038960a/appveyor.yml#L58). The problem is fixed in master for vmtk, so while we wait for a updated vmtk package to be relseased in conda we have added a warning to the installation instructions with a workaround for the problem.
