### 2018-12-12 4a2ff9d
Set smooth to default, and fix bug in get_line_to_change
