### 2018-12-12 1973815
Add manipulate_branch to setup
