### 2018-12-12 43b6d50
Remove unused global values
