### 2018-12-13 2bcda39
Improve method for moving points in Voronoi diagram, clean up common, split stenosis method into two methods, and add new method for changing area
