+++ morphman/manipulate_area.py
@@ -88,8 +88,7 @@ def manipulate_area(input_filepath, method, smooth, smooth_factor, no_smooth,
     centerline_regions = [centerline_splined, centerline_remaining]
     if centerline_diverging is not None:
         for div_cl in centerline_diverging:
-            centerline_diverging = [extract_single_line(div_cl, 0, startID=diverging_ids[0])]
-            centerline_regions += centerline_diverging
+            centerline_regions += [extract_single_line(div_cl, 0, startID=diverging_ids[0])]
     voronoi_regions = split_voronoi_with_centerlines(voronoi, centerline_regions)
 
     # Write the seperate segments
@@ -98,41 +97,13 @@ def manipulate_area(input_filepath, method, smooth, smooth_factor, no_smooth,
     for i in range(2, len(voronoi_regions)):
         write_polydata(voronoi_regions[i], voronoi_div_path.format(i-1))
 
-    # Compute area
+    # Compute area, and acount for diverging branches.
     if centerline_diverging is not None:
-        tmp_surface = create_new_surface(voronoi_regions[0],
+        surface_area = create_new_surface(voronoi_regions[0],
                                          poly_ball_size=poly_ball_size)
-        ###### FIXME
-        ###### For creating surface to plot
-        #from IPython import embed; embed()
-        center1, center2 = centerline_splined.GetPoint(0), centerline_splined.GetPoint(centerline_splined.GetNumberOfPoints() - 1)
-        normals = get_array("FrenetTangent", centerline_splined, k=3)
-        normal1, normal2 = normals[0], normals[-1]
-        plane1 = vtk_plane(center1, normal1)
-        plane2 = vtk_plane(center1, -normal1)
-
-        tmp_surface, tmp = clip_polydata(tmp_surface, plane1)
-        tmp_surface = attach_clipped_regions(tmp_surface, tmp, center1)
-        tmp_surface2, _ = clip_polydata(surface, plane2)
-        tmp_surface2 = get_connectivity(tmp_surface2, mode="Closest",
-                                        closest_point=center1)
-
-        plane1 = vtk_plane(center2, normal2)
-        plane2 = vtk_plane(center2, -normal2)
-        tmp_surface, tmp = clip_polydata(tmp_surface, plane2)
-        tmp_surface = attach_clipped_regions(tmp_surface, tmp, center2)
-        tmp_surface3, _ = clip_polydata(surface, plane1)
-        tmp_surface3 = get_connectivity(tmp_surface3, mode="Closest",
-                                        closest_point=center2)
-
-       ############
-
-        write_polydata(tmp_surface, surface_roi_path)
-        write_polydata(tmp_surface2, surface_roi_path.replace(".vtp", "2.vtp"))
-        write_polydata(tmp_surface3, surface_roi_path.replace(".vtp", "3.vtp"))
     else:
         surface_area = surface
-    centerline_area, centerline_area_sections = vmtk_compute_centerline_sections(surface,
+    centerline_area, centerline_area_sections = vmtk_compute_centerline_sections(surface_area,
                                                                                  centerline_splined)
     write_polydata(centerline_area, centerline_area_spline_path)
     write_polydata(centerline_area_sections, centerline_area_spline_sections_path)
@@ -212,17 +183,24 @@ def get_factor(line_to_change, method, beta, ratio, percentage, region_of_intere
             factor_ = ((area / mean_area) ** beta)[:, 0]
             factor = factor_ * (1 - trans) + trans
 
+    # Create or remove stenosis
     elif method == "stenosis":
-        if len(region_points) == 3:
-            t = np.linspace(0, np.pi, line_to_change.GetNumberOfPoints())
-            factor = (1 - np.sin(t) * percentage * 0.01).tolist()
-            factor = factor * (1 - trans) + trans
[…]
