### 2018-12-13 c8390d5
Update area for fixing bug and add test
