### 2018-12-15 2c5544f
Multiple bug fixes
