+++ morphman/manipulate_area.py
@@ -78,6 +78,7 @@ def manipulate_area(input_filepath, method, smooth, smooth_factor, no_smooth,
         centerline_diverging, region_points, diverging_ids = get_line_to_change(capped_surface, centerlines,
                                                                                 region_of_interest, method,
                                                                                 region_points, stenosis_length)
+
     write_polydata(centerline_splined, centerline_spline_path)
     write_polydata(centerline_remaining, centerline_remaining_path)
     if centerline_diverging is not None:
@@ -109,10 +110,11 @@ def manipulate_area(input_filepath, method, smooth, smooth_factor, no_smooth,
     write_polydata(centerline_area_sections, centerline_area_spline_sections_path)
 
     # Manipulate the Voronoi diagram
-    new_voronoi = change_area(voronoi_regions[0], centerline_area, method, beta, ratio, percentage,
-                              region_of_interest, region_points, centerline_diverging,
-                              voronoi_regions[2:])
-
+    factor = get_factor(centerline_area, method, beta, ratio, percentage,
+                        region_of_interest, region_points)
+    new_voronoi, new_centerlines = change_area(voronoi_regions[0], factor, centerline_area,
+                                               centerline_diverging, voronoi_regions[2:],
+                                               surface_area, centerlines)
     new_voronoi = merge_data([new_voronoi, voronoi_regions[1]])
     write_polydata(new_voronoi, voronoi_new_path)
 
@@ -121,8 +123,11 @@ def manipulate_area(input_filepath, method, smooth, smooth_factor, no_smooth,
     new_surface = create_new_surface(new_voronoi, poly_ball_size=poly_ball_size)
 
     print("-- Smoothing, clean, and check surface.")
-    new_surface = prepare_surface_output(new_surface, surface, centerlines,
-                                         output_filepath, test_merge=True)
+    changed = True if centerline_diverging is not None else False
+    new_surface = prepare_surface_output(new_surface, surface, new_centerlines,
+                                         output_filepath, changed=changed,
+                                         old_centerline=centerlines, test_merge=True)
+    write_polydata(new_centerlines, output_filepath.replace(".vtp", "_centerline.vtp"))
     write_polydata(new_surface, output_filepath)
 
 
@@ -207,9 +212,8 @@ def get_factor(line_to_change, method, beta, ratio, percentage, region_of_intere
     return factor
 
 
-def change_area(voronoi, line_to_change, method, beta, ratio, percentage,
-                region_of_interest, region_points, diverging_centerline,
-                diverging_voronoi):
+def change_area(voronoi, factor, line_to_change, diverging_centerline, diverging_voronoi,
+                surface_area, centerlines):
     """
     Change the cross-sectional area of an input
     voronoi diagram along the corresponding area
@@ -217,23 +221,16 @@ def change_area(voronoi, line_to_change, method, beta, ratio, percentage,
 
     Args:
         voronoi (vtkPolyData): Voronoi diagram.
+        factor (ndarray): An array with a factor for changing each point along the centerline
         line_to_change (vtkPolyData): Centerline representing area of interest.
-        method (str): Type of manipulation of the centerline.
-        beta (float): Factor deciding how area will change. Ignored if ratio is given.
-        ratio (float): Desired ratio between min and max cross-sectional area.
-        percentage (float): Percentage the area of the geometry / stenosis is increase/decreased.
[…]
