### 2018-12-17 4827af1
Bug fixes for area
