### 2018-12-17 bcbe8e9
Bug fix area
