### 2019-01-18 #1065 @brainstorm
I just gave it another look/try and I would advise against accepting the **installation** facet of this paper.

I still would recommend to have vmtk packaged in conda-forge (as a dependency), as @rlizzo diligently pointed out in https://github.com/vmtk/conda-recipes/issues/2 as well as in https://github.com/KVSlab/morphMan/issues/14.

From a functional perspective, I would like to have easier means to download test data locally in order to evaluate the software (had to fish it out from the testsuite). Something like a simple `scripts/morphman-data-download` or similar would help a lot getting newcomers acquainted with the software fast.

Other than those two small details, I would totally endorse this nice piece of software.
