### 2019-01-19 7bd6e04
Expose all functions in manipulate_* when importing morphman
