+++ morphman/__init__.py
@@ -1,5 +1,5 @@
 # Main features
-from .manipulate_area import manipulate_area, main_area
-from .manipulate_bifurcation import manipulate_bifurcation, main_bifurcation
-from .manipulate_bend import manipulate_bend, main_bend
-from .manipulate_curvature import manipulate_curvature, main_curvature
+from .manipulate_area import *
+from .manipulate_bifurcation import *
+from .manipulate_bend import *
+from .manipulate_curvature import *
