### 2019-01-19 90515f4
Rename read_command_line function and make it possible to parse inputfile and outputfile from a text, and return a dictionary with default values. This makes running morphman much easier when importing from python
