+++ morphman/manipulate_area.py
@@ -278,9 +278,15 @@ def change_area(voronoi, line_to_change, method, beta, ratio, percentage,
     return new_voronoi
 
 
-def read_command_line():
+def read_command_line_area(input_path=None, output_path=None):
     """
-    Read arguments from commandline
+    Read arguments from commandline and return all values in a dictionary.
+    If input_path and output_path are not None, then do not parse command line, but
+    only return default values.
+
+    Args:
+        input_path (str): Input file path, positional argument with default None.
+        output_path (str): Output file path, positional argument with default None.
     """
     # Description of the script
     description = "Manipulates the area of a tubular geometry. The script changes the area" + \
@@ -291,7 +297,8 @@ def read_command_line():
     parser = ArgumentParser(description=description, formatter_class=RawDescriptionHelpFormatter)
 
     # Add common arguments
-    add_common_arguments(parser)
+    required = not (input_path is not None and output_path is not None)
+    add_common_arguments(parser, required=required)
 
     # Mode / method for manipulation
     parser.add_argument("-m", "--method", type=str, default="variation",
@@ -353,8 +360,11 @@ def read_command_line():
                              " area of the geometry is increase/decreased overall or only" +
                              " stenosis")
 
-    # Parse
-    args = parser.parse_args()
+    # Parse paths to get default values
+    if required:
+        args = parser.parse_args()
+    else:
+        args = parser.parse_args(["-i" + input_path, "-o" + output_path])
 
     if args.method == "stenosis" and args.region_of_interest == "first_line":
         raise ValueError("Can not set region of interest to 'first_line' when creating or" +
