### 2019-01-19 12614fe
Convert all the turorials into additional tutorials
