### 2019-01-19 2f3e236
Make commandline variable 'size' consistent with variable name
