### 2019-01-19 623a97c
Fix bugs in demos
