### 2019-01-19 #1065 @aslakbergersen
Thank you for testing the software @brainstorm, it is much appreciated.

>  From a functional perspective, I would like to have easier means to download test data locally in order to evaluate the software (had to fish it out from the testsuite). Something like a simple scripts/morphman-data-download or similar would help a lot getting newcomers acquainted with the software fast.

I agree that there should be an additional script to ease playing around with the test data. To take it one step further, I took the command line examples from the Tutorial and rewrote them as python demos as well, see https://github.com/KVSlab/morphMan/pull/20. I also added a script which downloads the test data. Please also note that there are links for downloading the test data in the [tutorials](https://morphman.readthedocs.io/en/latest/getting_started.html#tutorials).

> I still would recommend to have vmtk packaged in conda-forge (as a dependency), as @rlizzo diligently pointed out in vmtk/conda-recipes#2 as well as in KVSlab/morphMan#14.

Just a clarification before starting to move the installation to conda-forge: Do I understand you correctly if you are asking us to add a _vmtk_ recipe to conda-forge, and then move _morphMan_ into conda-forge when it is accepted? If that is the case, I would like to confer with the main developers of _vmtk_ first, like @rlizzo. Or, do you propose that we add vmtk as an outside dependency, although [disadvised](https://conda-forge.org/docs/recipe.html#avoid-dependencies-outside-of-conda-forge).
