### 2019-01-19 #1065 @brainstorm
That's fantastic Aslak, thanks for putting it together as an script!

And yes, I would advise to add `vmtk` into conda-forge itself, it's relatively straightforward to add dependencies there [(did it myself in the past for packages for which I'm not author)](https://github.com/conda-forge/staged-recipes/pulls?q=is%3Apr+author%3Abrainstorm+is%3Aclosed). The main reason behind this preference is my experience with OSS projects in custom anaconda repositories not maintained by a bigger community (conda-forge).

When developers/PhDs/postdocs/people move on with other duties, if the dependencies are not maintained as well as the main package, [bitrot](https://en.wikipedia.org/wiki/Software_rot) happens faster, that's why I would like to insist on that point.
