### 2019-01-31 #1065 @aslakbergersen
@brainstorm, I agree that moving the project into conda-forge is the better option. We have been working on adding `vmtk` to conda-forge, but we are facing [some problems](https://github.com/conda-forge/staged-recipes/pull/7609) with getting it to build properly against `vtk` and `itk`. We will hopefully resolve this issue soon.

Adding `morphMan` to conda-corge will be much simpler since it is pure python.
