### 2019-02-16 #1065 @aslakbergersen
@trallard thank you for the follow-up. We are still working on moving `vmtk` over to conda-forge as requested by @brainstorm, such that also `morphMan` can be added to conda-forge. However, it has proven to be a bit challenging, and it is not finished yet, but is work in progress. To follow the progress, please see conda-forge/staged-recipes/pull/7609.
