### 2019-03-07 #1065 @aslakbergersen
@brainstorm, we have tried to add ´vmtk´ into ´conda-forge´, however without any luck conda-forge/staged-recipes/pull/7609. At this point, we, the vmtk developers, and members of conda-forge have all been unable to get any progress on this the last two weeks. We feel like we have exhausted all options on solving this problem within a reasonable timeframe.

Since ´vmtk´ is not our software, is already published in [JOSS](https://github.com/openjournals/joss-reviews/issues/745), and our installation methods are in satisfaction with JOSS criteria, we kindly ask this not to be a blocking issue for publication, but rather just a regular issue in the repository.

CC: @trallard
