### 2019-03-07 #1065 @brainstorm
@aslakbergersen Agreed, I've been following the threads on the side, thanks for putting the effort into it anyway. Hopefully upstream vmtk/conda will at some point in the future realise that the build/integration needs more love and fix it for good.

@trallard Go ahead, all clear from my side now after seeing this. Thanks everyone!
