### 2019-03-07 a80e0bd
Use same functions as in the other methods
