+++ morphman/manipulate_curvature.py
@@ -16,7 +16,7 @@ def manipulate_curvature(input_filepath, smooth, smooth_factor, smooth_factor_li
                          region_points, resampling_step, no_smooth, no_smooth_point):
     """
     Create a sharper or smoother version of the input geometry,
-    determined by a smoothed version of the siphon centerline.
+    determined by a smoothed version of the centerline.
 
     Args:
         input_filepath (str): Path to input surface.
@@ -40,6 +40,9 @@ def manipulate_curvature(input_filepath, smooth, smooth_factor, smooth_factor_li
     direction = "smoothed" if smooth_line else "extended"
     centerlines_path = base_path + "_centerline.vtp"
     centerline_smooth_path = base_path + "_centerline_smoothed.vtp"
+    centerline_spline_path = base_path + "_centerline_splined.vtp"
+    centerline_diverging_path = base_path + "_centerline_diverging.vtp"
+    centerline_remaining_path = base_path + "_centerline_remaining.vtp"
     new_centerlines_path = base_path + "_centerline_new_%s.vtp" % direction
 
     # Clean and capp / uncapp surface
@@ -48,7 +51,7 @@ def manipulate_curvature(input_filepath, smooth, smooth_factor, smooth_factor_li
     # Voronoi diagrams filenames
     voronoi_remaining_path = base_path + "_voronoi_curvature_remaining.vtp"
     voronoi_region_path = base_path + "_voronoi_region.vtp"
-    voronoi_diverging_path = base_path + "_voronoi_diverging.vtp"
+    voronoi_div_path = base_path + "_voronoi_diverging_{}.vtp"
 
     # Compute centerlines
     inlet, outlets = get_centers(surface, base_path)
@@ -60,78 +63,49 @@ def manipulate_curvature(input_filepath, smooth, smooth_factor, smooth_factor_li
                                           smooth, smooth_factor, no_smooth,
                                           no_smooth_point, voronoi, pole_ids)
     # Get region of interest
-    _, _, _, region_points, _ = get_line_to_change(capped_surface, centerlines,
-                                                region_of_interest, "variation", region_points, 0)
-    region_points = [[region_points[3 * i], region_points[3 * i + 1], region_points[3 * i + 2]]
-                     for i in range(len(region_points) // 3)]
+    centerline_splined, centerline_remaining, \
+            centerline_diverging, region_points, diverging_ids = get_line_to_change(capped_surface, centerlines,
+                                                                                    region_of_interest, "variation",
+                                                                                    region_points,
+                                                                                    None)
+    write_polydata(centerline_splined, centerline_spline_path)
+    write_polydata(centerline_remaining, centerline_remaining_path)
+    if centerline_diverging is not None:
+        write_polydata(merge_data(centerline_diverging), centerline_diverging_path)
 
-    # Set and get clipping points, centerlines and diverging centerlines
-    centerlines_complete, diverging_centerlines, region_points, region_points_vtk, diverging_ids = \
-        find_region_of_interest_and_diverging_centerlines(centerlines, region_points)
-    diverging_centerline_ispresent = diverging_centerlines is not None
-    diverging_id = None if len(diverging_ids) == 0 else diverging_ids[0]
+    # Split the Voronoi diagram
+    print("-- Clipping Voronoi diagram")
+    centerline_regions = [centerline_splined, centerline_remaining]
+    if centerline_diverging is not None:
+        for i, div_cl in enumerate(centerline_diverging):
+            centerline_regions += [extract_single_line(div_cl, 0,
[…]
