+++ morphman/common/common.py
@@ -13,9 +13,11 @@ import numpy as np
 import numpy.linalg as la
 import vtk
 from scipy.interpolate import splrep, splev
+from itertools import islice
 from scipy.signal import resample
 from vmtk import vtkvmtk, vmtkscripts
 from vtk.util import numpy_support
+from vtk.numpy_interface import dataset_adapter as dsa
 
 # Local import
 from morphman.common.vmtkpointselector import *
@@ -65,11 +67,25 @@ def read_polydata(filename, datatype=None):
         reader = vtk.vtkSTLReader()
         reader.MergingOn()
     elif fileType == 'vtk':
-        reader = vtk.vtkPolyDataReader()
+        # Read header
+        with open(filename) as myfile:
+            head = "".join(list(islice(myfile, 10))).lower()
+
+        # Set reader based on header
+        if "unstructured_grid" in head:
+            reader = vtk.vtkUnstructuredGridReader()
+        elif "structured_grid" in head:
+            reader = vtk.vtkStructuredGridReader()
+        elif "rectilinear_grid" in head:
+            reader = vtk.vtkRectilinearGridReader()
+        elif "structured_points" in head:
+            reader = vtk.vtkStructuredPointsReader()
+        elif "polydata" in head:
+            reader = vtk.vtkPolyDataReader()
     elif fileType == 'vtp':
         reader = vtk.vtkXMLPolyDataReader()
     elif fileType == 'vts':
-        reader = vtk.vtkXMinkorporereLStructuredGridReader()
+        reader = vtk.vtkXMLStructuredGridReader()
     elif fileType == 'vtr':
         reader = vtk.vtkXMLRectilinearGridReader()
     elif fileType == 'vtu':
@@ -747,6 +763,24 @@ def threshold(surface, name, lower=0, upper=1, threshold_type="between", source=
     return surface
 
 
+def compute_volume(surface):
+    """
+    Compute volume of closed surface
+
+    Args:
+        surface (vtkPolyData): Surfae to compute volume off
+
+    Returns:
+        volume (float): Volume of the input surface
+    """
+    mass_filter = vtk.vtkMassProperties()
+    mass_filter.SetInputData(surface)
+    mass_filter.Update()
+    volume = mass_filter.GetVolume()
+
+    return volume
+
+
 def compute_area(surface):
     """
     Compute area of polydata
@@ -759,8 +793,10 @@ def compute_area(surface):
     """
     mass = vtk.vtkMassProperties()
     mass.SetInputData(surface)
+    mass.Update()
+    area = mass.GetSurfaceArea()
 
-    return mass.GetSurfaceArea()
+    return area
 
 
 def clipp_capped_surface(surface, centerlines, clipspheres=0):
@@ -802,6 +838,31 @@ def clipp_capped_surface(surface, centerlines, clipspheres=0):
     return surface
 
 
+def compute_normals(surface, source="cell"):
+    """
+    A wrapper around vtkPolyDataNormals.
+
+    Args:
+        surface (vtkPolyData): Surface to generate normals from
+        source (str): Either cell or point
+
+    Returns:
+        surface (vtkPolyData): Surface with array 'Normals'
+    """
+    normal_generator = vtk.vtkPolyDataNormals()
+    normal_generator.SetInputData(surface)
+    if source == "cell":
+        normal_generator.ComputePointNormalsOff()
+        normal_generator.ComputeCellNormalsOn()
+    else:
+        normal_generator.ComputePointNormalsOn()
+        normal_generator.ComputeCellNormalsOff()
+
+    normal_generator.Update()
+
+    return normal_generator.GetOutput()
+
+
 def uncapp_surface(surface, gradients_limit=0.15, area_limit=0.3, circleness_limit=3):
     """
[…]
