### 2019-03-08 f9d1b78
Add parameters to splining method
