+++ morphman/manipulate_bifurcation.py
@@ -14,7 +14,7 @@ from morphman.common import *
 def manipulate_bifurcation(input_filepath, output_filepath, smooth, smooth_factor, angle,
                     keep_fixed_1, keep_fixed_2, bif, lower, no_smooth, no_smooth_point,
                     poly_ball_size, cylinder_factor, resampling_step,
-                    region_of_interest, region_points):
+                    region_of_interest, region_points, tension, continuity):
     """
     Objective rotation of daughter branches, by rotating
     centerlines and Voronoi diagram about the bifurcation center.
@@ -146,35 +146,43 @@ def manipulate_bifurcation(input_filepath, output_filepath, smooth, smooth_facto
     write_polydata(voronoi_clipped, voronoi_clipped_path)
 
     # Rotate branches (Centerline and Voronoi diagram)
-    print("-- Rotate centerlines and voronoi diagram.")
-    rotated_cl = rotate_cl(patch_cl, end_points[1], m, R)
-    write_polydata(rotated_cl, centerline_rotated_path)
+    angle = 0 if keep_fixed_1 and keep_fixed_2 else angle
+    if angle != 0:
+        print("-- Rotate centerlines and voronoi diagram.")
+        rotated_cl = rotate_cl(patch_cl, end_points[1], m, R)
+        write_polydata(rotated_cl, centerline_rotated_path)
 
-    if lower or bif:
-        rotated_bif_cl = rotate_cl(patch_bif_cl, end_points_bif[1], m, R)
-        write_polydata(rotated_bif_cl, centerline_rotated_bif_path)
+        if lower or bif:
+            rotated_bif_cl = rotate_cl(patch_bif_cl, end_points_bif[1], m, R)
+            write_polydata(rotated_bif_cl, centerline_rotated_bif_path)
 
-    rotated_voronoi = rotate_voronoi(voronoi_clipped, patch_cl, end_points[1], m, R)
-    write_polydata(rotated_voronoi, voronoi_rotated_path)
+        rotated_voronoi = rotate_voronoi(voronoi_clipped, patch_cl, end_points[1], m, R)
+        write_polydata(rotated_voronoi, voronoi_rotated_path)
+    else:
+        rotated_cl = patch_cl
+        rotated_voronoi = voronoi_clipped
+        if lower or bif:
+            rotated_bif_cl = patch_bif_cl
 
     # Interpolate the centerline
     print("-- Interpolate centerlines.")
     interpolated_cl = interpolate_patch_centerlines(rotated_cl, centerline_par, div_points[0].GetPoint(0),
-                                                    None, False)
+                                                    True, tension, continuity)
     write_polydata(interpolated_cl, centerline_new_path.replace(".vtp", "1.vtp"))
 
     if bif:
         interpolated_bif = interpolate_patch_centerlines(rotated_bif_cl, centerline_bif,
-                                                         None, "bif", True)
+                                                         None, True, tension, continuity)
         write_polydata(interpolated_bif, centerline_new_bif_path)
 
     if lower:
-        center = ((1 / 9.) * div_points[1][0] + (4 / 9.) * div_points[1][1] +
-                  (4 / 9.) * div_points[1][2]).tolist()
+        center = ((2 / 9.) * div_points[1][0] + (3.5 / 9.) * div_points[1][1] +
+                  (3.5 / 9.) * div_points[1][2]).tolist()
+        print("Center", center)
         div_points[0].SetPoint(0, center[0], center[1], center[2])
         interpolated_bif_lower = interpolate_patch_centerlines(rotated_bif_cl, centerline_bif,
                                                                div_points[0].GetPoint(0),
-                                                               "lower", True)
[…]
