### 2019-03-12 #1065 @trallard
Hi all thanks for the time spent in this review, it seems that all the issues have already been dealt with and there are just a few steps between acceptance and now.

@aslakbergersen  can you please make sure to complete the following tasks before completing this?

- [ ]   make a new release of the package (unless the latest release is up to date with the recent modifications) and add let us know what is the most up to date version
- [ ]  upload the revised software to your DOI-granting data/software repository
- [ ] post the DOI here
- [ ] double check the submission paper to ensure it is up-to-date (citations are correct, DOIs are added, etc)

Once these are completed I will process the last editorials tasks for acceptance of the software
