### 2019-03-12 cc8a991
Updated paper.lib with Doi's
