+++ paper/paper.bib
@@ -6,6 +6,7 @@
   Number                   = {9064},
   Pages                    = {1498 -- 1504},
   Volume                   = {349},
+  Doi                      = {10.1016/s0140-6736(96)07492-2},
   publisher={Elsevier}
 }
 
@@ -17,6 +18,7 @@
   number={16},
   pages={1829--1836},
   year={1996},
+  Doi                      = {10.1001/archinte.156.16.1829},
   publisher={American Medical Association}
 }
 
@@ -28,6 +30,8 @@
   Number                   = {21},
   Pages                    = {2035--2042},
   Volume                   = {282},
+  Doi                      = {10.1001/jama.282.21.2035},
+
 }
 
 @article{valen2018real,
@@ -36,6 +40,7 @@
   journal={Cardiovascular engineering and technology},
   pages={1--21},
   year={2018},
+  Doi                      = {10.1007/s13239-018-00374-2},
   publisher={Springer}
 }
 
@@ -45,6 +50,7 @@
   journal={Cardiovascular engineering and technology},
   pages={1--17},
   year={2018},
+  Doi                      = {10.1007/s13239-018-00376-0},
   publisher={Springer}
 }
 
@@ -55,7 +61,8 @@
   Year                     = {2016},
   Number                   = {3},
   Pages                    = {484--492},
-  Volume                   = {115},
+  Volume                   = {115},   
+  Doi                      = {10.1160/th15-07-0597},
   Publisher                = {Schattauer Publishers}
 }
 
@@ -66,6 +73,7 @@
   Year                     = {2013},
   Number                   = {12},
   Pages                    = {3561-3564},
+  Doi                      = {10.1161/strokeaha.113.001762},
   Volume                   = {44}
 }
 
@@ -91,6 +99,7 @@
   number={3},
   pages={1188--1203},
   year={2010},
+  Doi                      = {10.1007/s10439-010-9901-0},
   publisher={Springer}
 }
 
@@ -102,6 +111,7 @@
   number={1},
   pages={108--113},
   year={2004},
+  Doi                      = {10.3171/jns.2004.101.1.0108},
   publisher={Journal of Neurosurgery Publishing Group}
 }
 
@@ -109,17 +119,20 @@
   title={Proximal Parent Vessel Tapering is Associated With Aneurysm at the Middle Cerebral Artery Bifurcation},
   author={Lauric, Alexandra and Greim-Kuczewski, Kilian and Antonov, Alexei and Dardik, Gabriel and Magida, Jane K and Hippelheuser, James E and Kono, Kenichi and Malek, Adel M},
   journal={Neurosurgery},
+  Doi                      = {10.1093/neuros/nyy152},
   year={2018}
 }
 
 @mastersthesis{Bergersen2016,
   title={Investigating the Link Between Patient-specific Morphology and Hemodynamics: Implications for Aneurism Initiation?},
   author={Bergersen, Aslak Wigdahl},
+  Doi                      = {10.1371/journal.pcbi.1006510},
   year={2016}
 }
 
 @mastersthesis{kjeldsberg2018investigating,
   title={Investigating the Interaction Between Morphology of the Anterior Bend and Aneurysm Initiation},
   author={Kjeldsberg, Henrik Aasen},
+  Doi                      = {10.14236/ewic/hci2018.41},
   year={2018}
 }
