### 2019-03-12 cb7f040
Formatting in paper.bib.
