+++ paper/paper.bib
@@ -1,138 +1,137 @@
 @Article{Murray1997,
-  Title                    = {Alternative projections of mortality and disability by cause 1990-2020: Global Burden of Disease Study},
-  author={Murray, Christopher JL and Lopez, Alan D},
-  Journal                  = {The Lancet},
-  Year                     = {1997},
-  Number                   = {9064},
-  Pages                    = {1498 -- 1504},
-  Volume                   = {349},
-  Doi                      = {10.1016/s0140-6736(96)07492-2},
-  publisher={Elsevier}
+    Title = {Alternative projections of mortality and disability by cause 1990-2020: Global Burden of Disease Study},
+    author = {Murray, Christopher JL and Lopez, Alan D},
+    Journal = {The Lancet},
+    Year = {1997},
+    Number = {9064},
+    Pages = {1498 -- 1504},
+    Volume = {349},
+    Doi = {10.1016/s0140-6736(96)07492-2},
+    publisher = {Elsevier}
 }
 
 @article{gage1996effect,
-  title={The effect of stroke and stroke prophylaxis with aspirin or warfarin on quality of life},
-  author={Gage, Brian F and Cardinalli, Andria B and Owens, Douglas K},
-  journal={Archives of internal medicine},
-  volume={156},
-  number={16},
-  pages={1829--1836},
-  year={1996},
-  Doi                      = {10.1001/archinte.156.16.1829},
-  publisher={American Medical Association}
+    title = {The effect of stroke and stroke prophylaxis with aspirin or warfarin on quality of life},
+    author = {Gage, Brian F and Cardinalli, Andria B and Owens, Douglas K},
+    journal = {Archives of internal medicine},
+    volume = {156},
+    number = {16},
+    pages = {1829--1836},
+    year = {1996},
+    Doi = {10.1001/archinte.156.16.1829},
+    publisher = {American Medical Association}
 }
 
 @Article{Malek1999b,
-  Title                    = {{Hemodynamic Shear Stress and Its Role in Atherosclerosis}},
-  Author                   = {Malek, Adel M and Alper, Seth and Izumo, Seigo},
-  Journal                  = {JAMA},
-  Year                     = {1999},
-  Number                   = {21},
-  Pages                    = {2035--2042},
-  Volume                   = {282},
-  Doi                      = {10.1001/jama.282.21.2035},
-
+    Title = {{ Hemodynamic Shear Stress and Its Role in Atherosclerosis }},
+    Author = {Malek, Adel M and Alper, Seth and Izumo, Seigo},
+    Journal = {JAMA},
+    Year = {1999},
+    Number = {21},
+    Pages = {2035--2042},
+    Volume = {282},
+    Doi = {10.1001/jama.282.21.2035}
 }
 
 @article{valen2018real,
-  title={Real-world variability in the prediction of intracranial aneurysm wall shear stress: the 2015 International Aneurysm CFD Challenge},
-  author={Valen-Sendstad, Kristian and Bergersen, Aslak W and Shimogonya, Yuji and Goubergrits, Leonid and Bruening, Jan and Pallares, Jordi and Cito, Salvatore and Piskin, Senol and Pekkan, Kerem and Geers, Arjan J and others},
-  journal={Cardiovascular engineering and technology},
-  pages={1--21},
-  year={2018},
-  Doi                      = {10.1007/s13239-018-00374-2},
-  publisher={Springer}
+    title = {Real-world variability in the prediction of intracranial aneurysm wall shear stress: the 2015 International Aneurysm CFD Challenge},
[…]
