### 2019-03-12 2c41bd4
Activation of conda using powershell for conda activate
