### 2019-03-12 e0baf37
Reset shell after init powershell
