### 2019-03-12 87b019d
Updating yml scripts with conda 4.6 syntax.
