+++ appveyor.yml
@@ -52,10 +52,19 @@ test_script:
   #- "%CMD_IN_ENV% conda init powershell"
   - "%CMD_IN_ENV% conda config --set always_yes yes"
   - "%CMD_IN_ENV% conda config --add channels vmtk"
-  - "%CMD_IN_ENV% conda install pytest vtk vmtk scipy numpy"
+  - "%CMD_IN_ENV% conda config --add channels morphman"
+  #- "%CMD_IN_ENV% conda install pytest vtk vmtk scipy numpy"
   #- "%CMD_IN_ENV% conda build --python %APP_CONDA_PY% ."
 
-  
+# Set up environment
+  - "%CMD_IN_ENV% conda create -n morphman_env morphman pytest"
+
+# Activate environment
+  - "%CMD_IN_ENV% conda activate morphman_env"
+
+# Useful for debugging any issues with conda
+  - "%CMD_IN_ENV% conda info -a"
+
   # Search and replace in vmtkcenterlines.py for / to //
   - '%CMD_IN_ENV% sed -i "s/len(self.SourcePoints)\/3/len\(self.SourcePoints\)\/\/3/g" ..\\..\\Python%CONDA_PY%_64\\lib\\site-packages\\vmtk\\vmtkcenterlines.py'
   - '%CMD_IN_ENV% sed -i "s/len(self.TargetPoints)\/3/len\(self.TargetPoints\)\/\/3/g" ..\\..\\Python%CONDA_PY%_64\\lib\\site-packages\\vmtk\\vmtkcenterlines.py'
