### 2019-03-12 527476c
Revert travis yml, using miniconda 4.5 (latest)
