### 2019-03-12 ad80ca4
Removed conda from conda activate
