### 2019-03-12 97eea59
Vmtkcenterlines fix for windows.
