### 2019-03-12 6f3898d
Added install pytest with conda
