### 2019-03-12 0feaa56
Cleanup in appveyor script. Structured similarly as on travis.
