### 2019-03-12 9605b54
Added morphman to mock import in docs
