+++ docs/source/conf.py
@@ -87,7 +87,7 @@ language = None
 # directories to ignore when looking for source files.
 exclude_patterns = []
 
-autodoc_mock_imports = ["numpy", "scipy", "vtk", "vmtk"]
+autodoc_mock_imports = ["numpy", "scipy", "vtk", "vmtk", "morphman"]
 
 # The reST default role (used for this markup: `text`) to use for all
 # documents.
