### 2019-03-12 6538da9
[skip ci] Bumped release version in conf.py and conda yaml
