+++ docs/source/conf.py
@@ -66,9 +66,9 @@ author = u'Aslak W. Bergersen & Henrik A. Kjeldsberg'
 # built documents.
 #
 # The short X.Y version.
-version = u'0.1'
+version = u'0.2'
 # The full version, including alpha/beta/rc tags.
-release = u'0.1'
+release = u'0.2'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
