+++ conda-build/meta.yaml
@@ -1,6 +1,6 @@
 package:
   name: "morphman"
-  version: "0.1"
+  version: "0.2"
 
 source:
   path: ../
