### 2019-03-12 #1065 @hkjeldsberg
@trallard , the latest version has been released as morphMan v0.2 (master branch).
Additionally, the doi for the new release is https://doi.org/10.5281/zenodo.2591725
Me and @aslakbergersen are also happy with the submission paper.
