### 2019-03-14 603c01c
Tmp placing manipulate_branch on separate feature branch
