+++ morphman/manipulate_branch.py
@@ -1,495 +1,0 @@
-##   Copyright (c) Aslak W. Bergersen, Henrik A. Kjeldsberg. All rights reserved.
-##   See LICENSE file for details.
-
-##      This software is distributed WITHOUT ANY WARRANTY; without even 
-##      the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR 
-##      PURPOSE.  See the above copyright notices for more information.
-
-from argparse import ArgumentParser, RawDescriptionHelpFormatter
-
-# Local import
-from IPython import embed
-
-from morphman.common import *
-
-surfaceNormalsArrayName = 'SurfaceNormalArray'
-frenetTangentArrayName = 'FrenetTangent'
-
-
-def manipulate_branch(input_filepath, output_filepath, smooth, smooth_factor, region_of_interest, region_points,
-                      poly_ball_size, no_smooth, no_smooth_point, resampling_step):
-    """
-    Primary script for moving a selected part of any blood vessel.
-    Relies on an input centerline, a surface geometry of a 3D blood vessel network,
-    and the magnitude of movement (alpha, beta) in horizontal and vertical direction respectively.
-
-    Defines in- and out-put files, computes voronoi diagram, and
-    moves voronoi diagram in the horizontal direction based on the
-    definitions in "Investigating the Interaction Between Morphology
-    of the Anterior Bend and Aneurysm Initiation" (2018).
-    Proceeds by computing the corresponding centerline in the new
-    geometries, used for postprocessing, geometric analysis and
-    meshing.
-
-    Continuation in the move_vessel_vertically-method for vertical movement.
-
-    Args:
-        input_filepath (str): Path to input surface.
-        output_filepath (str): Path to output the manipulated surface.
-        smooth (bool): Smooth the Voronoi diagram.
-        smooth_factor (float): Smoothing factor used for Voronoi diagram smoothing.
-        region_of_interest (str): Method for setting the region of interest ['manual' | 'commandline' | 'landmarking']
-        region_points (list): If region_of_interest is 'commandline', this a flatten list of the start and endpoint
-        poly_ball_size (list): Resolution of polyballs used to create surface.
-        no_smooth (bool): True of part of the model is not to be smoothed.
-        no_smooth_point (ndarray): Point which is untouched by smoothing.
-        resampling_step (float): Resampling length when resampling centerline.
-    """
-
-    # Input filenames
-    base_path = get_path_names(input_filepath)
-
-    # Centerlines filenames
-    centerlines_path = base_path + "_centerline.vtp"
-    diverging_centerline_branch_path = base_path + "_centerlines_branch.vtp"
-
-    # Voronoi diagrams filenames
-    new_voronoi_path = base_path + "_voronoi_interpolated.vtp"
-
-    # Clean and capp / uncapp surface
-    surface, capped_surface = prepare_surface(base_path, input_filepath)
-
-    # Compute centerlines
-    inlet, outlets = get_centers(surface, base_path)
-
-    print("-- Compute centerlines and Voronoi diagram")
-    centerlines_complete, voronoi, pole_ids = compute_centerlines(inlet, outlets, centerlines_path,
-                                                                  capped_surface, resampling=resampling_step,
-                                                                  smooth=False, base_path=base_path)
-    if smooth:
-        voronoi = prepare_voronoi_diagram(capped_surface, centerlines_complete, base_path,
-                                          smooth, smooth_factor, no_smooth,
[…]
