### 2019-03-14 e2876e7
Moved initial manipulate_branch.py to separate branch. Update setup.py & version number
