### 2019-03-14 8585f5d
Refactoring in mani_branch. Added global params in common.
