+++ morphman/manipulate_branch.py
@@ -9,28 +9,45 @@ from argparse import ArgumentParser, RawDescriptionHelpFormatter
 
 from common import *
 
+
 # Local import
 
-surfaceNormalsArrayName = 'SurfaceNormalArray'
-frenetTangentArrayName = 'FrenetTangent'
+
+def test_data(capped_surface):
+    branch_point_surface_id = 12728
+    region_points = [[37.119407653808594, 27.06595230102539, 33.817745208740234],
+                     [37.82146072387695, 22.682409286499023, 28.8624267578125]]
+    branch_point = capped_surface.GetPoint(branch_point_surface_id)
+    return branch_point_surface_id, region_points, branch_point
+
+
+def get_new_branch_position(capped_surface):
+    print("\nPlease select point to move branch in the render window.")
+    seed_selector = vmtkPickPointSeedSelector()
+    seed_selector.SetSurface(capped_surface)
+    seed_selector.text = "Press space to select the point where you want to move the branch to, 'u' to undo.\n"
+    seed_selector.Execute()
+    new_branch_pos_id = seed_selector.GetTargetSeedIds().GetId(0)
+    new_branch_pos = capped_surface.GetPoint(new_branch_pos_id)
+
+    return new_branch_pos_id, new_branch_pos
 
 
 def manipulate_branch(input_filepath, output_filepath, smooth, smooth_factor, region_of_interest, region_points,
                       poly_ball_size, no_smooth, no_smooth_point, resampling_step):
     """
-    Primary script for moving a selected part of any blood vessel.
-    Relies on an input centerline, a surface geometry of a 3D blood vessel network,
-    and the magnitude of movement (alpha, beta) in horizontal and vertical direction respectively.
+    Primary script for moving a selected branch of any blood vessel.
+    Relies on a surface geometry of a 3D blood vessel network.
 
     Defines in- and out-put files, computes voronoi diagram, and
-    moves voronoi diagram in the horizontal direction based on the
-    definitions in "Investigating the Interaction Between Morphology
-    of the Anterior Bend and Aneurysm Initiation" (2018).
+    moves voronoi diagram according to user selected points.
+    Used selects two points defining a diverging branch, and the
+    one point defining the new location on the surface.
+    Moves, and performs rotation of the voronoi diagram, then
+    reconstructs the Voronoi diagram, creating a new surface.
     Proceeds by computing the corresponding centerline in the new
     geometries, used for postprocessing, geometric analysis and
     meshing.
-
-    Continuation in the move_vessel_vertically-method for vertical movement.
 
     Args:
         input_filepath (str): Path to input surface.
@@ -66,19 +83,13 @@ def manipulate_branch(input_filepath, output_filepath, smooth, smooth_factor, re
                                                                   capped_surface, resampling=resampling_step,
                                                                   smooth=False, base_path=base_path)
     if smooth:
+        print("-- Smoothing Voronoi diagram")
         voronoi = prepare_voronoi_diagram(capped_surface, centerlines_complete, base_path,
                                           smooth, smooth_factor, no_smooth,
                                           no_smooth_point, voronoi, pole_ids)
 
-        # FIXME: For testing only:
-        # write_polydata(centerlines_complete, "COMPLETE.vtp")
-        # branch_point = (52.46571350097656, 28.395702362060547, 17.509746551513672)
-
-    # region_points = np.asarray([[34.54340362548828, 7.8101677894592285, 43.735019683837899],
-    #                            [36.73946762084961, 28.79076385498047, 36.41799545288086]])
-    region_points = np.asarray([[38.820396423339844, 20.89110565185547, 28.940338134765625],
-                                [38.287437438964844, 27.811443328857422, 34.812583923339844]])
[…]
