### 2019-03-18 #1065 @aslakbergersen
@trallard, is the information provided by @hkjeldsberg what you needed for starting the final steps of publication? Unfortunately, we do not have the rights to "check off" the task-boxes in your post, but I think we have addressed all of them.
