### 2019-03-19 46d3a7d
Fix bug in defining diverging centerlines
