+++ morphman/common/common.py
@@ -2958,7 +2958,12 @@ def get_line_to_change(surface, centerline, region_of_interest, method, region_p
         id2 = locator.FindClosestPoint(end_point)
         p1_tmp = line.GetPoint(id1)
         p2_tmp = line.GetPoint(id2)
-        if distance(start_point, p1_tmp) < tol and distance(end_point, p2_tmp) < tol:
+        close_start = distance(start_point, p1_tmp) < tol
+        close_end = distance(end_point, p2_tmp) < tol
+
+        # Check if the centerline is going through both or None of the start and end of
+        # the region of interest
+        if close_start == close_end:
             if start_id != 0:
                 tmp = extract_single_line(centerline, i, startID=0, endID=start_id - 1)
                 remaining_centerlines.append(tmp)
