### 2019-03-19 c11af29
Additional if test for defining diverging centerlines
