### 2019-03-20 #1065 @trallard
@aslakbergersen  and @hkjeldsberg all looks good to me now so I am going to proceed for acceptance ⚡✨

@brainstorm and @rlizzo thank you so much for your time and expertise as reviewers it is much appreciated at JOSS 🙌🏻

@openjournals/joss-eics this submission has been accepted and is ready to be published 🎉🙌🏻🥳
