### 2019-03-20 c47c5b5
Update description of area and make sure that nknots is smaler than the number of points in the centerline to be splines
