### 2019-03-20 #1065 @arfon
@aslakbergersen @hkjeldsberg - could you merge this PR: https://github.com/KVSlab/morphMan/pull/23 - it will improve the formatting of your citations when there are multiples.
