### 2019-03-20 #1065 @aslakbergersen
@arfon Thank you for the PR, it is now merged.
