### 2019-03-20 #1065 @arfon
> pandoc: 10.21105.joss.01065.pdf: openBinaryFile: does not exist (No such file or directory)

Sometimes @whedon times out generating these PDFs...
